% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{aio_collect}
\alias{aio_collect}
\alias{aio_collect_}
\title{Collect the Data of a List of Aios}
\usage{
aio_collect(x)

aio_collect_(x)
}
\arguments{
\item{x}{a list of Aios (objects of class \sQuote{sendAio}, \sQuote{recvAio}
or \sQuote{ncurlAio}).}
}
\value{
A list of objects (the same length as \sQuote{x}, preserving names).
}
\description{
\code{aio_collect} collects the data of a list of Aios in a list, waiting for
    resolution if still in progress.

\code{aio_collect_} is a variant that allows user interrupts, suitable for
    interactive use.
}
\details{
This function will wait for the asynchronous operation(s) to
    complete if still in progress (blocking).
}
\seealso{
\code{\link{aio_data}} for retrieving the data of a single Aio.
}
