% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncurl.R
\name{as.promise.asyncUrl}
\alias{as.promise.asyncUrl}
\title{Make asyncUrl Promise}
\usage{
\method{as.promise}{asyncUrl}(x)
}
\arguments{
\item{x}{an object of class \sQuote{asyncUrl}.}
}
\value{
A \sQuote{promise} object.
}
\description{
Creates a \sQuote{promise} from an \sQuote{asyncUrl} object.
}
\details{
This function is an S3 method for the generic \code{as.promise} for
    class \sQuote{asyncUrl}.

    Requires the \pkg{promises} package.

    Allows an \sQuote{asyncUrl} to be used with the promise pipe
    \code{\%...>\%}, which schedules a function to run upon resolution of the
    Aio.
}
