% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises.R
\name{as.promise.ncurlAio}
\alias{as.promise.ncurlAio}
\title{Make ncurl Promise}
\usage{
\method{as.promise}{ncurlAio}(x)
}
\arguments{
\item{x}{an object of class \sQuote{ncurlAio}.}
}
\value{
A \sQuote{promise} object.
}
\description{
Creates a \sQuote{promise} from an \sQuote{ncurlAio} object.
}
\details{
This function is an S3 method for the generic \code{as.promise} for
    class \sQuote{ncurlAio}.

    Requires the \CRANpkg{promises} package.

    Allows an \sQuote{ncurlAio} to be used with the promise pipe
    \code{\%...>\%}, which schedules a function to run upon resolution of the
    Aio.
}
\examples{
if (interactive() && requireNamespace("promises", quietly = TRUE)) {

library(promises)

p <- as.promise(ncurl_aio("https://www.cam.ac.uk/"))
print(p)
is.promise(p)

p2 <- ncurl_aio("https://postman-echo.com/get") \%...>\% identity()
p2$then(cat)
is.promise(p2)

}

}
