% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{as.promise.recvAio}
\alias{as.promise.recvAio}
\title{Make recvAio Promise}
\usage{
\method{as.promise}{recvAio}(x)
}
\arguments{
\item{x}{an object of class 'recvAio'.}
}
\value{
A 'promise' object.
}
\description{
Creates a 'promise' from an 'recvAio' object.
}
\details{
This function is an S3 method for the generic \code{as.promise} for class
'recvAio'.

Requires the \pkg{promises} package.

Allows a 'recvAio' to be used with the promise pipe \verb{\%...>\%}, which schedules
a function to run upon resolution of the Aio.
}
