% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{close}
\alias{close}
\alias{close.nanoSocket}
\alias{close.nanoContext}
\alias{close.nanoDialer}
\alias{close.nanoListener}
\title{Close Connection}
\usage{
\method{close}{nanoSocket}(con, quietly = TRUE, ...)

\method{close}{nanoContext}(con, quietly = TRUE, ...)

\method{close}{nanoDialer}(con, quietly = TRUE, ...)

\method{close}{nanoListener}(con, quietly = TRUE, ...)
}
\arguments{
\item{con}{a Socket, Context, Dialer or Listener.}

\item{quietly}{[default TRUE] if FALSE, confirmation that the object has been
successfully closed is printed to the console (stdout), useful for logging
purposes.}

\item{...}{not used.}
}
\value{
Zero (invisibly) on success.
}
\description{
Close Connection on a Socket, Context, Dialer or Listener.
}
\details{
Closing an object explicitly frees its resources. An object can also
    be removed directly in which case its resources are freed when the object
    is garbage collected.

    Dialers and Listeners are implicitly closed when the socket they are
    associated with is closed.

    Closing a socket associated with a context also closes the context.

    When closing a socket or a context: messages that have been submitted for
    sending may be flushed or delivered, depending upon the transport. Closing
    the socket while data is in transmission will likely lead to loss of that
    data. There is no automatic linger or flush to ensure that the socket
    send buffers have completely transmitted.
}
