% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{collect_pipe}
\alias{collect_pipe}
\title{Collect the Pipe from an Aio}
\usage{
collect_pipe(x)
}
\arguments{
\item{x}{a 'recvAio' object.}
}
\value{
A Pipe (object of class \sQuote{nanoPipe}).
}
\description{
This function blocks until it has received a message from the Aio. The Pipe
    used to receive the message is returned and the message is still
    available for retrieval by normal means. A Pipe is a low-level object and
    it is not normally necessary to deal with them directly.
}
\details{
As Pipes are always owned by a Socket, removing (and garbage
    collecting) a Pipe does not close it or free its resources. A Pipe may,
    however, be explicitly closed.
}
\examples{
s <- socket("rep")
listen(s)
s1 <- socket("req")
dial(s1)
r <- recv_aio(s, timeout = 500)

if (!send(s1, "")) {
  p <- tryCatch(collect_pipe(r), error = identity)
  print(p)
  reap(p)
}

close(s)
close(s1)

}
