% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{Open Context}
\usage{
context(socket)
}
\arguments{
\item{socket}{a Socket.}
}
\value{
A new Context (object of class 'nanoContext' and 'nano').
}
\description{
Open a new Context to be used with a Socket. The purpose of a Context is to
    permit applications to share a single socket, with its underlying dialers
    and listeners, while still benefiting from separate state tracking.
}
\details{
Contexts allow the independent and concurrent use of stateful
    operations using the same socket. For example, two different contexts
    created on a rep socket can each receive requests, and send replies to
    them, without any regard to or interference with each other.

    Note: only certain protocols support creation of separate contexts -
    namely req, rep, surveyor, respondent.

    To send and receive over a context use \code{\link{send}} and
    \code{\link{recv}} or their async counterparts \code{\link{send_aio}} and
    \code{\link{recv_aio}}.

    For nano objects, use the \code{$context()} method, which will return a
    new context.
}
\examples{
s <- socket("req", listen = "inproc://nanonext")
ctx <- context(s)
ctx
close(ctx)
close(s)

n <- nano("req", listen = "inproc://nanonext")
ctx <- n$context()
ctx
close(ctx)
n$close()

}
