% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{cv_flag}
\alias{cv_flag}
\title{Condition Variable Flag Value}
\usage{
cv_flag(cv)
}
\arguments{
\item{cv}{a \sQuote{conditionVariable} object.}
}
\value{
An integer.
}
\description{
Query the flag value of a conditionVariable.
}
\keyword{internal}
