% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_new}
\alias{cv_new}
\title{New Condition Variable}
\usage{
cv_new()
}
\value{
A 'conditionVariable' object.
}
\description{
Creates a new condition variable (protected by a mutex). Pass this to the
    special forms of the functions returning 'recvAio' objects which also
    take a condition variable object and signals this when the receive is
    complete.
}
\examples{
cv <- cv_new()

}
