% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_reset}
\alias{cv_reset}
\title{Reset a Condition Variable}
\usage{
cv_reset(cv)
}
\arguments{
\item{cv}{a 'conditionVariable' object.}
}
\value{
Invisible NULL.
}
\description{
Resets the internal condition for a condition variable to an inital zero value.
}
\examples{
cv <- cv_new()
cv_reset(cv)

}
