% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{cv_value}
\alias{cv_value}
\alias{cv_reset}
\alias{cv_signal}
\title{Condition Variables - Utilities}
\usage{
cv_value(cv)

cv_reset(cv)

cv_signal(cv)
}
\arguments{
\item{cv}{a 'conditionVariable' object.}
}
\value{
For \strong{cv_value}: integer value of the condition variable.

    For \strong{cv_reset} and \strong{cv_signal}: zero (invisibly).
}
\description{
\code{cv_value} inspects the internal value of a condition variable.

\code{cv_reset} resets the internal value and flag of a condition variable.

\code{cv_signal} signals a condition variable.
}
\details{
These functions work with a 'conditionVariable'. See \code{\link{cv}}
    for further details of how the condition and flag operates.
}
\examples{
cv <- cv()

cv_value(cv)

cv_reset(cv)

cv_signal(cv)
cv_value(cv)

}
