% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{device}
\alias{device}
\title{Create Device}
\usage{
device(s1, s2)
}
\arguments{
\item{s1}{a raw mode Socket.}

\item{s2}{a raw mode Socket.}
}
\value{
Invisibly, an integer exit code. If the device was successfully
    created, this function does not return.
}
\description{
Creates a device which is a socket forwarder or proxy. Provides for improved
    horizontal scalability, reliability, and isolation.
}
\details{
Only raw mode sockets may be used with this function. Sockets s1 and
    s2 must be compatible with each other, i.e. be opposite halves of a two
    protocol pattern, or both the same protocol for a single protocol pattern.
}
\section{Usage}{


    Warning: this function is designed to be called in an isolated process
    with the two sockets. Once called, it will block with no ability to
    interrupt. Kill the process to terminate the device.
}

