% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.active}
\alias{.active}
\title{Read Active Status}
\usage{
.active(sock)
}
\arguments{
\item{sock}{a dispatcher Socket.}
}
\value{
An vector of integer values.
}
\description{
Reads the active status of threaded dispatcher sockets (for internal use
    only).
}
\keyword{internal}
