% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.advance}
\alias{.advance}
\title{Advances the RNG State}
\usage{
.advance()
}
\value{
NULL.
}
\description{
Internal package function.
}
\examples{
.Random.seed
invisible(.advance())
.Random.seed

}
\keyword{internal}
