% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.dispatcher}
\alias{.dispatcher}
\title{Dispatcher Socket}
\usage{
.dispatcher(host, url, tls = NULL)
}
\arguments{
\item{host}{\sQuote{inproc://} url connecting the host to the thread.}

\item{url}{the URLs at which to listen for rep nodes.}

\item{tls}{[default NULL] for secure tls+tcp:// or wss:// connections only,
provide a TLS configuration object created by \code{\link{tls_config}}.}
}
\value{
A \sQuote{req} Socket. The thread is attached as an attribute.
}
\description{
Creates a Dispatcher socket, which is a special type of \sQuote{req} socket,
    with FIFO scheduling using a threaded implementation (for internal use
    only).
}
\keyword{internal}
