% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\name{.dispatcher_stop}
\alias{.dispatcher_stop}
\title{Stop In-Process Dispatcher}
\usage{
.dispatcher_stop(disp)
}
\arguments{
\item{disp}{External pointer to dispatcher handle.}
}
\value{
Invisible NULL.
}
\description{
Stop In-Process Dispatcher
}
\keyword{internal}
