% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\name{.dispatcher_wait_n}
\alias{.dispatcher_wait_n}
\title{Wait for N Daemon Connections}
\usage{
.dispatcher_wait_n(disp, n)
}
\arguments{
\item{disp}{External pointer to dispatcher handle.}

\item{n}{Number of connections to wait for.}
}
\value{
Invisible NULL.
}
\description{
Wait for N Daemon Connections
}
\keyword{internal}
