% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.header}
\alias{.header}
\title{Set Serialization Header}
\usage{
.header(x = 0L)
}
\arguments{
\item{x}{integer value.}
}
\value{
The integer value \code{x} supplied.
}
\description{
Internal package function.
}
\keyword{internal}
