% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.interrupt}
\alias{.interrupt}
\title{Interrupt Switch}
\usage{
.interrupt()
}
\value{
NULL.
}
\description{
Toggles on or off whether async receive completions trigger an interrupt.
Internal package function.
}
\examples{
.interrupt()
.interrupt()

}
\keyword{internal}
