% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\name{.limit_gate}
\alias{.limit_gate}
\title{Limit Gate}
\usage{
.limit_gate(disp)
}
\arguments{
\item{disp}{External pointer to dispatcher handle.}
}
\value{
Invisible NULL.
}
\description{
Block until inflight count is below limit, then increment.
}
\keyword{internal}
