% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\name{.limit_release}
\alias{.limit_release}
\title{Limit Release}
\usage{
.limit_release(disp)
}
\arguments{
\item{disp}{External pointer to dispatcher handle.}
}
\value{
Invisible NULL.
}
\description{
Decrement inflight count on request failure.
}
\keyword{internal}
