% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mark}
\alias{.mark}
\title{Set Serialization Marker}
\usage{
.mark(bool = TRUE)
}
\arguments{
\item{bool}{logical value.}
}
\value{
The logical \code{bool} supplied.
}
\description{
Internal package function.
}
\keyword{internal}
