% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.online}
\alias{.online}
\title{Read Online Status}
\usage{
.online(sock)
}
\arguments{
\item{sock}{a dispatcher Socket.}
}
\value{
An vector of integer values.
}
\description{
Reads the online status of threaded dispatcher sockets (for internal use
only).
}
\keyword{internal}
