% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{.promise}
\alias{.promise}
\title{Create Promise}
\usage{
.promise(x, ctx)
}
\arguments{
\item{x}{a \sQuote{recvAio} or \sQuote{ncurlAio} object.}

\item{ctx}{the return value of \sQuote{environment()}.}
}
\value{
NULL.
}
\description{
Internal package function.
}
\details{
If successful, both \sQuote{x} and \sQuote{ctx} are preserved and
    accessible from the promise callback.
}
\keyword{internal}
