% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rng_adv}
\alias{.rng_adv}
\title{Advances the RNG State}
\usage{
.rng_adv()
}
\value{
An empty list.
}
\description{
Internal package function.
}
\examples{
.Random.seed
invisible(.rng_adv())
.Random.seed

}
\keyword{internal}
