% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{.unresolved}
\alias{.unresolved}
\title{Technical Utility: Query if an Aio is Unresolved}
\usage{
.unresolved(aio)
}
\arguments{
\item{aio}{an Aio (object of class 'sendAio' or 'recvAio').}
}
\value{
Logical TRUE if 'aio' is an unresolved Aio, or FALSE otherwise.
}
\description{
Query whether an Aio or Aio value remains unresolved. This function is a fast
    alternative to \code{\link{unresolved}} not intended for ordinary use.
    Simply queries the busy status of an Aio and does not alter its state in
    any way i.e. does not attempt to retrieve any result status or messages.
}
