% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.until}
\alias{.until}
\title{Condition Variables - Until}
\usage{
.until(cv, msec)
}
\arguments{
\item{cv}{a 'conditionVariable' object.}

\item{msec}{maximum time in milliseconds to wait for the condition variable
to be signalled.}
}
\value{
(invisibly) logical TRUE if signalled, or else FALSE if the timeout
    was reached.
}
\description{
This function is identical to \code{\link{until}} and deprecated.
}
