% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{format_sse}
\alias{format_sse}
\title{Format Server-Sent Event}
\usage{
format_sse(data, event = NULL, id = NULL, retry = NULL)
}
\arguments{
\item{data}{The data payload. Will be prefixed with "data: " on each line.}

\item{event}{[default NULL] Optional event type (e.g., "message", "error").}

\item{id}{[default NULL] Optional event ID for client reconnection.}

\item{retry}{[default NULL] Optional retry interval in milliseconds.}
}
\value{
A character string formatted as an SSE message.
}
\description{
Formats a message according to the SSE specification.
}
\details{
SSE messages have this format:

\if{html}{\out{<div class="sourceCode">}}\preformatted{event: <event-type>
id: <event-id>
retry: <milliseconds>
data: <data-line-1>
data: <data-line-2>

}\if{html}{\out{</div>}}

Each message ends with two newlines. Multi-line data is split and each
line prefixed with "data: ".
}
\examples{
format_sse(data = "Hello")
#> "data: Hello\n\n"

format_sse(data = "Hello", event = "greeting")
#> "event: greeting\ndata: Hello\n\n"

format_sse(data = "Line 1\nLine 2")
#> "data: Line 1\ndata: Line 2\n\n"

}
