% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{gen_key}
\alias{gen_key}
\title{Generate Key}
\usage{
gen_key(filename = "keyfile.key")
}
\arguments{
\item{filename}{[default 'keyfile.key'] character file path / name at which
to store the generated key.}
}
\value{
Invisibly, 'filename'.
}
\description{
Generate a 4096 bit RSA private/public key pair in the PEM format using Mbed
    TLS; the first step in creating a self-signed certificate.
}
\examples{

if (interactive()) {
# Only run examples in interactive R sessions

keyfile <- gen_key(filename = tempfile())
key <- readLines(keyfile)

unlink(keyfile)
key

}

}
\keyword{internal}
