% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{getopt}
\alias{getopt}
\title{Get Option for a Socket, Context, Stream, Listener or Dialer}
\usage{
getopt(object, type = c("bool", "int", "ms", "size", "string", "uint64"), opt)
}
\arguments{
\item{object}{a Socket, Context, Stream, Listener or Dialer.}

\item{type}{[default 'bool'] type of option - either 'bool', 'int', 'ms'
(duration), 'size', 'string' or 'uint64'. Alternatively, for performance,
specify an integer position in the vector of choices e.g. 1L for 'bool',
2L for 'int' etc.}

\item{opt}{name of option, e.g. 'reconnect-time-min', as a character string.
See \link{opts}.}
}
\value{
The value of the option (logical for type 'bool', integer for 'int',
    'ms' and 'size', character for 'string' and double for 'uint64') if
    successful, or else an integer 'errorValue'.
}
\description{
Get value of \link{opts} for a Socket, Context, Stream, Listener or Dialer.
}
\details{
To get options for a Listener or Dialer attached to a Socket or nano
    object, pass in the objects directly via for example \code{$listener[[1]]}
    for the first Listener.
}
\examples{
s <- socket("pair")
getopt(s, "int", "send-buffer")
close(s)

s <- socket("req")
ctx <- context(s)
getopt(ctx, "ms", "send-timeout")
close(ctx)
close(s)

s <- socket("pair", dial = "inproc://nanonext", autostart = FALSE)
getopt(s$dialer[[1]], "ms", "reconnect-time-min")
close(s)

s <- socket("pair", listen = "inproc://nanonext", autostart = FALSE)
getopt(s$listener[[1]], "size", "recv-size-max")
close(s)

}
