% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>>\%}
\alias{\%>>\%}
\title{Deferred Evaluation Pipe}
\usage{
x \%>>\% f
}
\arguments{
\item{x}{a value that is possibly an 'unresolvedValue'.}

\item{f}{a function that accepts 'x' as its first argument.}
}
\value{
The evaluated result, or if x is an 'unresolvedValue', an
    'unresolvedExpr' encapsulating the eventual evaluation result. Query its
    \code{$data} field for resolution.
}
\description{
Pipe a possibly unresolved value forward into a function. [experimental]
}
\details{
Supports stringing together a series of piped expressions (as per
    the below example).

    This function is marked [experimental], which means it is currently
    under development. Please note that the final implementation is likely to
    differ from the current version.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

s1 <- socket("pair", listen = "inproc://nanonext")
s2 <- socket("pair", dial = "inproc://nanonext")

msg <- recv_aio(s2)
b <- msg$data \%>>\% c(2, 3) \%>>\% as.character()
b
res <- send_aio(s1, 1)
b$data

close(s1)
close(s2)
}

}
