% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ip_addr}
\alias{ip_addr}
\title{IP Address}
\usage{
ip_addr()
}
\value{
A character string.
}
\description{
Returns the local network IPv4 address, or an empty character string if no
appropriate network adaptors were found.
}
\examples{
ip_addr()

}
