% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nano}
\alias{is_nano}
\title{Is Nano}
\usage{
is_nano(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE or FALSE.
}
\description{
Is the object an object created by \{nanonext\} i.e. a nanoSocket,
    nanoContext, nanoStream, nanoListener, nanoDialer or nano Object.
}
\details{
Note: does not include Aio objects, for which there is a separate
    function \code{\link{is_aio}}.
}
\examples{
s <- socket()
is_nano(s)
n <- nano()
is_nano(n)

close(s)
n$close()

}
