% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nul_byte}
\alias{is_nul_byte}
\title{Is Nul Byte}
\usage{
is_nul_byte(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE or FALSE.
}
\description{
Is the object a nul byte.
}
\examples{
is_nul_byte(as.raw(0L))
is_nul_byte(raw(length = 1L))
is_nul_byte(writeBin("", con = raw()))

is_nul_byte(0L)
is_nul_byte(NULL)
is_nul_byte(NA)

}
