% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{lock}
\alias{lock}
\alias{unlock}
\title{Lock / Unlock a Socket}
\usage{
lock(socket, cv = NULL, decr = FALSE)

unlock(socket)
}
\arguments{
\item{socket}{a Socket.}

\item{cv}{(optional) a 'conditionVariable'. If supplied, the socket is locked
only while the value of the condition variable is non-zero.}

\item{decr}{[default FALSE] applicable only if 'cv' is specified, should be
set to TRUE only if a \code{\link{pipe_notify}} event on removal has been
registered on 'cv'. In this case, the counter within the
'conditionVariable' is decremented to offset the increment that is
registered for the pipe removal, even though this occurs before it has
been added to the socket.}
}
\value{
Invisibly, zero on success (will otherwise error).
}
\description{
Prevents further pipe connections from being established at a Socket. If a
    socket is locked, new pipe connections are closed before they can be
    added to the socket.
}
\examples{
s <- socket("bus", listen = "inproc://nanolock")
s1 <- socket("bus", dial = "inproc://nanolock")
lock(s)
s2 <- socket("bus", dial = "inproc://nanolock")

send(s, "test")
recv(s1)
recv(s2)

unlock(s)
s3 <- socket("bus", dial = "inproc://nanolock")
send(s, "test")
recv(s1)
recv(s3)

close(s)
close(s1)
close(s2)
close(s3)

}
