% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mclock}
\alias{mclock}
\title{Clock Utility}
\usage{
mclock()
}
\value{
A double.
}
\description{
Provides the number of elapsed milliseconds since an arbitrary reference time
    in the past. The reference time will be the same for a given program, but
    may differ between programs.
}
\details{
A convenience function for building concurrent applications. The
    resolution of the clock depends on the underlying system timing facilities
    and may not be particularly fine-grained. This utility should however be
    faster than using \code{Sys.time()}.
}
\examples{
time <- mclock(); msleep(100); mclock() - time

}
