% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messenger.R
\name{messenger}
\alias{messenger}
\title{Messenger}
\usage{
messenger(url)
}
\arguments{
\item{url}{a URL to connect to, specifying the transport and address as
a character string e.g. 'tcp://127.0.0.1:5555' (see \link{transports}).}
}
\value{
Invisible NULL.
}
\description{
Multi-threaded console-based 2-way messaging system based on NNG scalability
    protocols.
}
\section{Usage}{


    Type outgoing messages and hit return to send.

    The timestamps of outgoing messages are prefixed by \code{>} and that of
    incoming messages by \code{<}.

    \code{:q} is the command to quit.

    NOTE: This is an experimental feature that is currently undergoing
    testing.
}

