% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_pipe}
\alias{msg_pipe}
\title{Message Pipe}
\usage{
msg_pipe(aio)
}
\arguments{
\item{aio}{a 'recvAio' object, not including 'ncurlAio' objects.}
}
\value{
A 'connectionPipe' object.
}
\description{
Returns the pipe associated with a received message. A pipe can be thought of
    as a single connection.
}
