% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nano_init}
\alias{nano_init}
\title{nanonext Initialise}
\usage{
nano_init(warn = c("immediate", "deferred", "error", "none"))
}
\arguments{
\item{warn}{[default 'immediate'] character string defining how to treat
warnings generated by the package. 'immediate' to print warnings as they
occur, 'deferred' to print warnings when evaluation returns to the top
level, 'error' to upgrade all warnings to errors (stops execution), and
'none' to ignore all warnings.}
}
\value{
Invisible NULL.
}
\description{
Initialise global options [DEPRECATED]. This function prints a deprecation
    message and no longer has any effect.
}
