% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nanonext_version}
\alias{nanonext_version}
\title{nanonext Version Information}
\usage{
nanonext_version()
}
\value{
A character vector of length one.
}
\description{
Provides the package version, NNG and mbed TLS library versions in a single
    string.
}
\examples{
nanonext_version()

}
