% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nng_clock}
\alias{nng_clock}
\title{NNG Clock Utility}
\usage{
nng_clock()
}
\value{
A double.
}
\description{
Provides the number of elapsed milliseconds since an arbitrary reference time
    in the past. The reference time will be the same for a given program, but
    may differ between programs.
}
\details{
A convenience function for building concurrent applications. The
    resolution of the clock depends on the underlying system timing facilities
    and may not be particularly fine-grained. This utility should however be
    faster than using base \code{Sys.time()}.
}
\examples{
time <- nng_clock(); Sys.sleep(0.1); nng_clock() - time

}
