% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nng_sleep}
\alias{nng_sleep}
\title{NNG Sleep Utility}
\usage{
nng_sleep(msec)
}
\arguments{
\item{msec}{integer number of milliseconds to block the caller.}
}
\value{
Invisible NULL.
}
\description{
Sleep function. May block for longer than requested, with the actual wait
    time determined by the capabilities of the underlying system.
}
\examples{
time <- nng_clock(); nng_sleep(100); nng_clock() - time

}
