% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{pipe_id}
\alias{pipe_id}
\title{Get the Pipe ID of a recvAio}
\usage{
pipe_id(x)
}
\arguments{
\item{x}{a resolved 'recvAio'.}
}
\value{
Integer pipe ID.
}
\description{
Caution: must only be used on an already-resolved 'recvAio'. This function
does not perform validation of these pre-conditions.
}
