% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{pipe_monitor}
\alias{pipe_monitor}
\alias{read_monitor}
\title{Create a Pipe Monitor}
\usage{
pipe_monitor(sock, cv, n)

read_monitor(x)
}
\arguments{
\item{sock}{a Socket.}

\item{cv}{a conditionVariable.}

\item{n}{initial size of monitor (number of pipes).}

\item{x}{an external pointer to a pipe monitor.}
}
\value{
An external pointer that may be passed to \code{read_monitor}.
}
\description{
This function monitors pipe additions and removals from a socket.
}
\examples{
cv <- cv()
s <- socket("rep", listen = "inproc://nanonext")
m <- pipe_monitor(s, cv, 8L)
read_monitor(m)
close(s)

}
