% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{pipes}
\alias{pipes}
\title{Obtain Open Pipes from a Socket}
\usage{
pipes(socket, n = 1L)
}
\arguments{
\item{socket}{a Socket.}

\item{n}{integer number of Pipes to return.}
}
\value{
A list of Pipes (objects of class \sQuote{nanoPipe}).
}
\description{
This function blocks until it has received \sQuote{n} messages from the
    supplied Socket. The Pipes used to receive those messages are returned in
    a list. A Pipe is a low-level object and it is not normally necessary to
    deal with them directly.
}
\details{
As Pipes are owned by the corresponding Socket, removing (and
    garbage collecting) a Pipe does not close it or free its resources. A
    Pipe may, however, be explicitly closed.
}
\examples{
s <- socket("rep")
listen(s)
s1 <- socket("req")
dial(s1)

if (!send(s1, "")) {
  p <- pipes(s, n = 1)
  print(p)
  lapply(p, close)
}

close(s)
close(s1)

}
