% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler_start}
\alias{profiler_start}
\title{Start the `gperftools` profiler.}
\usage{
profiler_start(path)
}
\arguments{
\item{path}{Character of length 1 with the file path to the profiling samples.}
}
\value{
`NULL` (invisibly).
}
\description{
Start the `gperftools` profiler.
}
