% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random}
\alias{random}
\title{NNG Random Number Generator}
\usage{
random(n = 1L)
}
\arguments{
\item{n}{[default 1L] length of vector to return.}
}
\value{
A length 'n' vector of random integers.
}
\description{
Strictly not for statistical analysis. Not reproducible and with no guarantee
    on desirable statistical properties. Provides random numbers suitable for
    system functions such as cryptographic key generation. Random values are
    obtained using platform-specific strong cryptographic random number
    facilities where available.
}
\details{
If 'n' is non-integer, it will be coerced to integer; if a vector,
    only the first element will be used.
}
\note{
This is an independent source of random numbers and does not affect the
    RNG state of the pseudo-random number generators present in R. The
    statistical properties of the outputs may not be desirable and hence
    should not be used in statistical analyses.
}
\examples{
random()
random(n = 3L)

}
