% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_stdin}
\alias{read_stdin}
\title{Read stdin}
\usage{
read_stdin()
}
\value{
a Socket.
}
\description{
Reads \code{stdin} from a background thread, allowing the stream to be accessed as
messages from an NNG 'inproc' socket. As the read is blocking, it can only be
used in non-interactive sessions. Closing \code{stdin} causes the background
thread to exit and the socket connection to end.
}
\details{
A 'pull' protocol socket is returned, and hence can only be used with receive
functions.
}
