% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{reap}
\alias{reap}
\title{Reap}
\usage{
reap(con)
}
\arguments{
\item{con}{a Socket, Context, Listener or Dialer.}
}
\value{
Integer zero, or else an integer 'errorValue' if an invalid object
    was supplied.
}
\description{
An alternative to \code{close} for Sockets, Contexts, Listeners and Dialers
    that runs finalisers, immediately freeing associated resources. For
    advanced use only.
}
\details{
May be used on unclassed external pointers e.g. those created by
    \code{\link{.context}}. Returns silently and does not warn or error. On
    success, the supplied object is stripped to a nil external pointer and
    may no longer be used.
}
\examples{
s <- socket("req")
listen(s)
dial(s)
ctx <- .context(s)

reap(ctx)
reap(s[["dialer"]][[1]])
reap(s[["listener"]][[1]])
reap(s)
reap(s)

}
\seealso{
\code{\link{close}}
}
