% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refhook}
\alias{refhook}
\title{Set and Retrieve Refhook Function}
\usage{
refhook(x = list())
}
\arguments{
\item{x}{a function or base64 encoded function, or NULL to reset.}
}
\value{
The current refhook function stored internally.
}
\description{
Sets the refhook function stored internally, used in all message send and
    receive operations. Calling without an argument retrieves the stored value.
}
\details{
The provided function maps to the 'refhook' argument of
    \code{\link{serialize}} or \code{\link{unserialize}}, as the case may be.
    As the same function is used in both cases, it should contain the logic
    to distinguish between them if necessary.
}
\examples{
refhook()

refhook(function(x) if (is_aio(x)) base64enc(x) else
    if (is.character(x)) unserialize(base64dec(x, convert = FALSE)))

refhook(NULL)

}
