% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{run_event_loop}
\alias{run_event_loop}
\title{Run the Event Loop}
\usage{
run_event_loop(timeout = Inf)
}
\arguments{
\item{timeout}{maximum time to wait in seconds. The default \code{Inf} blocks
until an event occurs.}
}
\value{
Logical \code{TRUE} if a callback was executed, \code{FALSE} otherwise.
}
\description{
Runs the \pkg{later} event loop, blocking until the next event and
dispatching any scheduled callbacks. Used to process a server's (see
\code{\link[=http_server]{http_server()}}) or other asynchronous operations' callbacks on R's main
thread.
}
\details{
Requires the \pkg{later} package.
}
