% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{self_signed_cert}
\alias{self_signed_cert}
\title{Generate Self-Signed Certificate}
\usage{
self_signed_cert(cn = "localhost", valid = "20301231235959")
}
\arguments{
\item{cn}{[default 'localhost'] character issuer common name (CN) for the
certificate. This can be either a hostname or an IP address, but must
match the actual server URL as client authentication will depend on it.}

\item{valid}{[default '20301231235959'] character 'not after' date-time in
'yyyymmddhhmmss' format. The certificate is not valid after this time.}
}
\value{
A list of length 2, comprising \code{$server} and \code{$client}.
    These may be passed directly to the relevant argument of
    \code{\link{tls_config}}.
}
\description{
Generate a self-signed x509 certificate from a 4096 bit RSA private/public
    key pair for use with encrypted TLS communications.
}
\examples{

if (interactive()) {
# Only run examples in interactive R sessions

cert <- self_signed_cert(cn = "127.0.0.1", valid = "20301231235959")
ser <- tls_config(server = cert$server)
cli <- tls_config(client = cert$client)

s <- socket()
s1 <- socket()
listen(s, "wss://127.0.0.1:5555", tls = ser)
dial(s1, "wss://127.0.0.1:5555", tls = cli)

# secure TLS connection established

close(s1)
close(s)

cert

}

}
