% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendrecv.R
\name{send_vec}
\alias{send_vec}
\title{Send Vector}
\usage{
send_vec(socket, data, block = FALSE, echo = TRUE)
}
\arguments{
\item{socket}{a Socket.}

\item{data}{a vector. Will be converted and sent as raw unless already a raw
vector in which case it will be sent unmodified.}

\item{block}{[default FALSE] logical flag whether to block until successful
or return immediately even if unsuccessful (e.g. no connection available).}

\item{echo}{[default TRUE] logical flag whether to return the raw vector of
sent data. Set to FALSE for performance-critical applications where
invisble NULL will be returned instead.}
}
\value{
Raw vector of sent data, or invisible NULL if 'echo' is set to FALSE.
}
\description{
Send vector data over a Socket. Data will be sent as binary without R
    serialisation, hence appropriate for interfacing with external programs.
}
\examples{
pub <- socket("pub", dial = "inproc://nanonext")

send_vec(pub, c(10.1, 20.2, 30.3))

close(pub)

}
