% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncurl.R
\name{server}
\alias{server}
\title{Start REST Server}
\usage{
server(url = "http://127.0.0.1:5555/api/rest")
}
\arguments{
\item{url}{full http address including hostname, port and path at which to
host the server.}
}
\value{
This function never returns.
}
\description{
Creates an instance of an HTTP REST server which evaluates R expressions sent
    to it [EXPERIMENTAL]. As arbitrary R expressions are evaluated, this
    should only be deployed on the local machine (using the 127.0.0.1
    loopback address) in a trusted environment.
}
\details{
Query the API with an HTTP client using the \sQuote{POST} method,
    with the request data being the R expression as a text string. The
    received response body will consist of the serialized evaluation result.
    Unserialize to return an R object.

    Use only in a new session. Use \sQuote{ctrl + \\} to forcibly quit
    when finished as the function blocks with no means of interruption.

    Currently still experimental as the server lacks error handling. Sending
    an invalid R expression will cause the server to exit.
}
\examples{
if (interactive()) {

# run in a new session:
# nanonext::server()

res <- ncurl("http://127.0.0.1:5555/api/rest",
             convert = FALSE,
             method = "POST",
             data = "Sys.time()")

if (!is_error_value(res$data)) unserialize(res$data)

}

}
