% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{sha1}
\alias{sha1}
\title{Cryptographic Hashing Using the SHA-1 Algorithm}
\usage{
sha1(x, key = NULL, convert = TRUE)
}
\arguments{
\item{x}{an object.}

\item{key}{(optional) supply a secret key to generate an HMAC. If missing or
NULL, the SHA-1 hash of 'x' is returned.}

\item{convert}{[default TRUE] logical value whether to convert the output to
a character string or keep as a raw vector.}
}
\value{
A raw vector or character string depending on 'convert', of byte
    length 20.
}
\description{
Returns a SHA-1 hash or HMAC of the supplied R object. Uses the optimised
    implementation from the Mbed TLS library. For secure applications, one of
    the SHA-2 algorithms such as \code{\link{sha256}} should be considered
    instead.
}
\details{
For arguments 'x' and 'key', a scalar string or raw vector (with no
    attributes) is hashed directly. All other objects are first serialised
    using R serialisation version 3, big-endian representation, with the
    serialization header stripped (for portability across R installations as
    this contains the R version number).

    The result of hashing is always a byte sequence, which is converted to a
    character string hex representation if 'convert' is TRUE, or returned as
    a raw vector if 'convert' is FALSE.
}
\examples{
# SHA-1 hash as character string:
sha1("hello world!")

# SHA-1 hash as raw vector:
sha1("hello world!", convert = FALSE)

# Obtain HMAC:
sha1("hello world!", "SECRET_KEY")

}
