% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{socket_pipe}
\alias{socket_pipe}
\title{Create a Pipe from a Pipe ID}
\usage{
socket_pipe(x)
}
\arguments{
\item{x}{integer pipe ID.}
}
\value{
A Pipe (object of class \sQuote{nanoPipe}).
}
\description{
This function creates a Pipe from the pipe identifier recorded at $aio in a
resolved Aio, for use with compatible functions such as \code{\link{send_aio}}.
A Pipe is a low-level object and it is not normally necessary to deal with
them directly.
}
\details{
As Pipes are always owned by a Socket, removing (and garbage collecting) a
Pipe does not close it or free its resources. A Pipe may, however, be
explicitly closed.
}
\examples{
s <- socket("rep", listen = "inproc://nanonext")
s1 <- socket("req", dial = "inproc://nanonext")

r <- recv_aio(s, timeout = 500)

if (!send(s1, "")) {
  call_aio(r)
  p <- socket_pipe(r$aio)
  print(p)
  reap(p)
}

close(s)
close(s1)

}
