% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{status_code}
\alias{status_code}
\title{Translate HTTP Status Codes}
\usage{
status_code(x)
}
\arguments{
\item{x}{numeric HTTP status code to translate.}
}
\value{
A character vector.
}
\description{
Translate HTTP response status codes, in the range 100 to 599, to human
    readable form. If the status code is not recognised (could be a custom
    code used by the server), 'Non-standard Response' is returned.
}
\examples{
status_code(200)
status_code(404)

}
