% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{stop_request}
\alias{stop_request}
\title{Stop Request Operation}
\usage{
stop_request(x)
}
\value{
Invisibly, loigcal \code{TRUE} or \code{FALSE}.
}
\description{
Stop an asynchronous Aio operation, or a list of Aio operations, created by
\code{\link[=request]{request()}}. This is an augmented version of \code{\link[=stop_aio]{stop_aio()}} that additionally
requests cancellation by sending an integer zero followed by the context ID
over the context, and waiting for the response.
}
\keyword{internal}
