% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strcat}
\alias{strcat}
\title{Concatenate 2 Strings}
\usage{
strcat(a, b)
}
\arguments{
\item{a}{character value.}

\item{b}{character value.}
}
\value{
A character vector of length 1.
}
\description{
A fast implementation that simply combines two strings into one.
}
\details{
If either 'a' or 'b' is a vector of length greater than 1, only the
    first element of each is concatenated.
}
\examples{
strcat("hello ", "world!")

}
