% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream}
\alias{stream}
\title{Open Stream}
\usage{
stream(dial = NULL, listen = NULL, textframes = FALSE)
}
\arguments{
\item{dial}{a URL to dial, specifying the transport and address as a character
string e.g. 'ipc:///tmp/anyvalue' or 'tcp://127.0.0.1:5555'
(not all transports are supported).}

\item{listen}{a URL to listen at, specifying the transport and address as a
character string e.g. 'ipc:///tmp/anyvalue' or 'tcp://127.0.0.1:5555'
(not all transports are supported).}

\item{textframes}{[default FALSE] applicable to the websocket transport only,
enables sending and receiving of TEXT frames (ignored otherwise).}
}
\value{
A Stream (object of class 'nanoStream' and 'nano').
}
\description{
Open a Stream by either dialing (establishing an outgoing connection) or
    listening (accepting an incoming connection) at an address. This is a
    low-level interface intended for communicating with non-NNG
    endpoints.
}
\details{
A Stream is used for raw byte stream connections. Byte streams are
    reliable in that data will not be delivered out of order, or with portions
    missing.

    Specify only one of 'dial' or 'listen'. If both are specified, 'listen'
    will be ignored.

    TLS is automatically configured for dialing a secure websocket address
    starting 'wss://' (where the NNG library has been built with TLS support).
}
