% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream_send_aio}
\alias{stream_send_aio}
\title{Send via Stream (Async)}
\usage{
stream_send_aio(stream, data, timeout)
}
\arguments{
\item{stream}{a Stream.}

\item{data}{a vector.}

\item{timeout}{(optional) in milliseconds.}
}
\value{
A 'sendAio' (object of class 'sendAio').
}
\description{
Send data asynchronously over a byte stream. This is a low-level interface
    intended for communicating with non-NNG endpoints [experimental].
}
\details{
Sending a byte stream asynchronously is non-blocking and returns
    a 'sendAio' immediately.

    For a 'sendAio', the send result is available at \code{$result}. An
    'unresolved' logical NA is returned if the async operation is yet to
    complete, The resolved value will be zero on success, or else an integer
    error code.

    To wait for and check the result of the send operation, use
    \code{\link{call_aio}} on the returned 'sendAio' object.

    Alternatively, to stop the async operation, use \code{\link{stop_aio}}.
}
