% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{tls_config}
\alias{tls_config}
\title{Create TLS Configuration}
\usage{
tls_config(client = NULL, server = NULL, pass = NULL, auth = is.null(server))
}
\arguments{
\item{client}{(optional) for creating a client configuration, the absolute
path to a file containing X.509 certificate(s) in PEM format, comprising
the certificate authority certificate chain (and revocation list if
present), used to validate certificates presented by peers.}

\item{server}{(optional) for creating a server configuration, the absolute
path to a single file containing the PEM encoded certificate and
associated private key, along with any additional certificates leading to
a validation chain, the leaf certificate first. It is not required to
include the self-signed root.}

\item{pass}{[default NULL] required only if the secret key contained in the
file supplied to 'server' is encrypted with a password. Do not provide
directly in case it is cached or recorded in history, but through an
object or function that returns the value.}

\item{auth}{(optional) logical value whether to require authentication - by
default TRUE for client and FALSE for server configurations. If TRUE, the
session is only allowed to proceed if the peer has presented a valid
certificate. If FALSE, authentication is optional, whereby a certificate
is validated if presented by the peer, but the session still allowed to
proceed if not. If neither 'client' nor 'server' are supplied, then no
authentication is performed and this argument has no effect. Supplying a
non-logical value will error.}
}
\value{
A 'tlsConfig' object.
}
\description{
Create a TLS configuration object to be used for secure connections.
}
\details{
Specify one of 'client' or 'server' only, or neither (in which case
    an empty client configuration is created), as a configuration can only be
    of one type.

    For use with \code{\link{ncurl}}, up-to-date CA certificates in PEM
    format, extracted from Mozilla, are available at:
    \url{https://curl.se/docs/caextract.html}. This link is not endorsed; use
    at your own risk.
}
\examples{
tls <- tls_config()
tls
ncurl("https://www.r-project.org/", timeout = 1000L, tls = tls)

}
