% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{until}
\alias{until}
\title{Wait Upon a Condition Variable Until a Certain Time}
\usage{
until(cv, msec)
}
\arguments{
\item{cv}{a 'conditionVariable' object.}

\item{msec}{maximum time in milliseconds to wait for the condition variable
to be signalled.}
}
\value{
Invisible NULL.
}
\description{
Waits upon a condition variable to be signalled by an incoming message being
    received, until a certain time has elapsed.
}
\details{
If 'msec' is non-integer, it will be coerced to integer. Non-numeric
    input will be ignored and return immediately.

    The internal condition maintains a state. Each signal increments the
    condition by 1. Each time this function returns due to being woken (as
    opposed to timeout) the condition is decremented by 1.
}
\examples{
cv <- cv_new()
until(cv, 10L)

}
