% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{wait}
\alias{wait}
\alias{until}
\alias{.until}
\title{Wait / Until}
\usage{
wait(cv)

until(cv, msec)

.until(cv, msec)
}
\arguments{
\item{cv}{a 'conditionVariable' object.}

\item{msec}{maximum time in milliseconds to wait for the condition variable
to be signalled.}
}
\value{
For \strong{wait} and \strong{until}: (invisibly) logical TRUE, or
    else FALSE if a flag has been set.

    For \strong{.until}: (invisibly) logical TRUE if signalled, or else FALSE
    if the timeout was reached.
}
\description{
\code{wait} waits on a condition being signalled by completion of an
    asynchronous receive.

\code{until} waits until a future time on a condition being signalled by
    completion of an asynchronous receive.
}
\details{
These synchronisation primitives take a 'conditionVariable'. See
    \code{\link{cv}} for further details of how the condition and flag
    operates.
}
\examples{
cv <- cv()

# wait(cv) # uncommenting will block until the cv is signalled

until(cv, 10L)

}
