% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{wait_aio}
\alias{wait_aio}
\title{Wait for the Value of an Asynchronous Aio Operation}
\usage{
wait_aio(aio)
}
\arguments{
\item{aio}{an Aio (object of class 'sendAio', 'recvAio' or 'ncurlAio').}
}
\value{
The passed object (invisibly).
}
\description{
Retrieve the value of an asynchronous Aio operation, waiting for the
    operation to complete if still in progress.
}
\details{
This function is identical to \code{\link{call_aio}} except that it
    is user-interruptible. If interrupted, the aio is stopped upon the next
    garbage collection event, and hence may return an 'errorValue' 20
    'Operation canceled' if it remains unresolved by that time.
}
\examples{
s1 <- socket("pair", listen = "inproc://nanonext")
s2 <- socket("pair", dial = "inproc://nanonext")

res <- send_aio(s1, data.frame(a = 1, b = 2), timeout = 100)
res
wait_aio(res)
res$result

msg <- recv_aio(s2, timeout = 100)
msg
wait_aio(msg)$data

close(s1)
close(s2)

}
