% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{with_lock}
\alias{with_lock}
\title{With Lock}
\usage{
with_lock(cv, expr)
}
\arguments{
\item{cv}{a \sQuote{conditionVariable} object.}

\item{expr}{an expression to evaluate under lock.}
}
\value{
The return value of \sQuote{expr}.
}
\description{
Evaluate an expression whilst holding a mutex lock internal to a
\sQuote{conditionVariable}. The purpose of this function is to prevent
access to objects guarded by the same lock from another thread.
}
\details{
Do not attempt to call any function using the conditionVariable \sQuote{cv}
in \sQuote{expr} otherwise you will create a deadlock.
}
\examples{
cv <- cv()
with_lock(cv, "hello")

}
