% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writec}
\alias{writec}
\title{Write C}
\usage{
writec(x)
}
\arguments{
\item{x}{character string.}
}
\value{
Invisible NULL. As a side effect, \code{x} is output to \code{stdout}.
}
\description{
Writes using the C function \code{write()} for non-buffered I/O. Avoids
interleaved output when writing concurrently to \code{stdout} from multiple
processes.
}
\details{
This function writes to the C-level \code{stdout} of the process and hence cannot
be re-directed by \code{\link[=sink]{sink()}}.
}
\examples{
writec("\n")

}
