% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zzz}
\alias{zzz}
\title{Internal Package Function}
\usage{
zzz()
}
\description{
Do not use. Only present for cleaning up after running examples and tests.
}
\examples{
rm(list = ls())
gc()
Sys.sleep(1L)
nanonext:::zzz()

}
\keyword{internal}
