% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendrecv.R
\name{aio_call}
\alias{aio_call}
\title{Call the Result of an Asynchronous AIO Operation}
\usage{
aio_call(aio)
}
\arguments{
\item{aio}{An Aio (object of class 'sendAio' or 'recvAio').}
}
\value{
For a 'sendAio', zero on success. For a 'recvAio', either a named list
    of 2 elements: 'raw' containing the received raw vector and 'data' containing
    the converted R object, or else just the converted R object depending on
    the parameters set.
}
\description{
Retrieve the result of an asynchronous AIO operation. Will wait for the AIO
    operation (blocking) if not yet complete. Once the result is retrieved,
    the Aio is deallocated and further actions cannot be performed on it.
}
\examples{
s1 <- socket("pair", listen = "inproc://nanonext")
s2 <- socket("pair", dial = "inproc://nanonext")

status <- send_aio(s1, data.frame(a = 1, b = 2), timeout = 100)
status
aio_call(status)

res <- recv_aio(s2, timeout = 100, keep.raw = FALSE)
res
aio_call(res)

close(s1)
close(s2)

}
