% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{aio_collect}
\alias{aio_collect}
\alias{aio_collect_}
\title{Collect Data of an Aio or List of Aios}
\usage{
aio_collect(x)

aio_collect_(x)
}
\arguments{
\item{x}{an Aio or list of Aios (objects of class \sQuote{sendAio},
\sQuote{recvAio} or \sQuote{ncurlAio}).}
}
\value{
Depending on the type of \sQuote{x} supplied, an object or list of
    objects (the same length as \sQuote{x}, preserving names).
}
\description{
\code{aio_collect} collects the data of an Aio or list of Aios, waiting for
    resolution if still in progress.

\code{aio_collect_} is a variant that allows user interrupts, suitable for
    interactive use.
}
\details{
This function will wait for the asynchronous operation(s) to
    complete if still in progress (blocking).
}
\examples{
s1 <- socket("pair", listen = "inproc://nanonext")
s2 <- socket("pair", dial = "inproc://nanonext")

res <- send_aio(s1, data.frame(a = 1, b = 2), timeout = 100)
aio_collect(res)

msg <- recv_aio(s2, timeout = 100)
aio_collect_(res)

close(s1)
close(s2)

}
