% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{aio_data}
\alias{aio_data}
\alias{aio_data_}
\title{Retrieve Data of an Aio or List of Aios}
\usage{
aio_data(x)

aio_data_(x)
}
\arguments{
\item{x}{an Aio or list of Aios (objects of class \sQuote{sendAio},
\sQuote{recvAio} or \sQuote{ncurlAio}).}
}
\value{
Depending on the type of \sQuote{x} supplied, either an object or a
    list of objects (the same length as \sQuote{x}, preserving names).
}
\description{
\code{aio_data} retrieves the data value of an Aio, or list of Aios, waiting
    for resolution if still in progress.

\code{aio_data_} is a variant that allows user interrupts, suitable for
    interactive use.
}
\details{
This function will wait for the asynchronous operation(s) to
    complete if still in progress (blocking).
}
