% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncurl.R
\name{as.promise.ncurlAio}
\alias{as.promise.ncurlAio}
\title{Make ncurlAio Promise}
\usage{
\method{as.promise}{ncurlAio}(x)
}
\arguments{
\item{x}{an object of class \sQuote{ncurlAio}.}
}
\value{
A \sQuote{promise} object.
}
\description{
Creates a \sQuote{promise} from an \sQuote{ncurlAio} object.
}
\details{
This function is an S3 method for the generic \code{as.promise} for class
\sQuote{ncurlAio}.

Requires the \pkg{promises} package.

Allows an \sQuote{ncurlAio} to be used with the promise pipe \verb{\%...>\%}, which
schedules a function to run upon resolution of the Aio.
}
