% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{as.promise.recvAio}
\alias{as.promise.recvAio}
\title{Make recvAio Promise}
\usage{
\method{as.promise}{recvAio}(x)
}
\arguments{
\item{x}{an object of class \sQuote{recvAio}.}
}
\value{
A \sQuote{promise} object.
}
\description{
Creates a \sQuote{promise} from an \sQuote{recvAio} object.
}
\details{
This function is an S3 method for the generic \code{as.promise} for class
\sQuote{recvAio}.

Requires the \pkg{promises} package.

Allows an \sQuote{recvAio} to be used with the promise pipe \code{\%...>\%},
which schedules a function to run upon resolution of the Aio.
}
