% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{call_aio}
\alias{call_aio}
\title{Call the Result of an Asynchronous AIO Operation}
\usage{
call_aio(aio, block = TRUE)
}
\arguments{
\item{aio}{An Aio (object of class 'sendAio' or 'recvAio').}

\item{block}{[default TRUE] whether to wait for completion of the AIO
operation (blocking) or return immediately.}
}
\value{
The passed Aio object (invisibly), or NULL if non-blocking and the
    Aio has yet to resolve.
}
\description{
Retrieve the result of an asynchronous AIO operation. Optionally wait for the
    AIO operation to complete if this is still in progress.
}
\details{
For a 'recvAio', the received raw vector will be attached in \code{$raw}
    (unless 'keep.raw' was set to FALSE when receiving), and the converted R
    object in \code{$data}.

    For a 'sendAio', the send result will be attached to the Aio in \code{$result}.
    This will be zero on success.

    To access the values directly, use for example on a sendAio 'x':
    \code{call_aio(x)$result}.

    For a 'recvAio', in case of an error in unserialisation or data conversion,
    the received raw vector will always be saved in \code{$raw} to allow the
    data to be recovered.

    Once the result has been successfully retrieved, the Aio is deallocated
    and only the result is stored in the Aio object.
}
\section{Non-blocking}{


    To query whether Aio \code{x} has resolved, test if \code{call_aio(x, block = FALSE)}
    returns NULL. When the Aio resolves, the Aio itself will be returned
    (invisibly) instead of NULL. The data may then be extracted from the Aio
    using \code{$result}, \code{$raw} or \code{$data} as the case may be.

    It is not advisable to try to extract the data from a 'recvAio' in one
    step using something like \code{call_aio(x)$data} in the non-blocking case.
    This is as \code{call_aio()} will return NULL if the Aio is unresolved and
    \code{NULL$data} is also \code{NULL}, hence it would be impossible to
    distinguish between an unresolved Aio and a NULL return value.
}

\examples{
s1 <- socket("pair", listen = "inproc://nanonext")
s2 <- socket("pair", dial = "inproc://nanonext")

res <- send_aio(s1, data.frame(a = 1, b = 2), timeout = 100)
res
call_aio(res)
res
res$result

res <- recv_aio(s2, timeout = 100)
res
call_aio(res)$data
res

close(s1)
close(s2)

}
