% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{cert_write}
\alias{cert_write}
\title{Write Certificate}
\usage{
cert_write(
  key = "keyfile.key",
  cn = "localhost",
  valid = "20301231235959",
  filename = "cert.crt"
)
}
\arguments{
\item{key}{[default 'keyfile.key'] character file path / name to the issuer
key file (as generated by \code{\link{gen_key}}).}

\item{cn}{[default 'localhost'] character issuer common name (CN) for the
certificate. This can be either a hostname or an IP address, but must
match the actual server URL as client authentication will depend on it.}

\item{valid}{[default '20301231235959'] character 'not after' date-time in
'yyyymmddhhmmss' format. The certificate is not valid after this time.}

\item{filename}{[default 'cert.crt'] character file path / name at which
to store the generated certificate.}
}
\value{
Invisibly, 'filename'.
}
\description{
Generate a self-signed certificate for use with encrypted TLS communications.
}
\examples{

if (interactive()) {
# Only run examples in interactive R sessions

keyfile <- gen_key(filename = tempfile())
certfile <- cert_write(key = keyfile, cn = "127.0.0.1", filename = tempfile())
cert <- readLines(certfile)

unlink(certfile)
unlink(keyfile)
cert

}

}
