% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{close}
\alias{close}
\alias{close.nanoSocket}
\alias{close.nanoContext}
\alias{close.nanoDialer}
\alias{close.nanoListener}
\title{Close Connection}
\usage{
\method{close}{nanoSocket}(con, ...)

\method{close}{nanoContext}(con, ...)

\method{close}{nanoDialer}(con, ...)

\method{close}{nanoListener}(con, ...)
}
\arguments{
\item{con}{a Socket, Context, Dialer or Listener.}

\item{...}{not used.}
}
\value{
Zero (invisibly) on success.
}
\description{
Close Connection on a Socket, Context, Dialer or Listener.
}
