% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{collect_aio}
\alias{collect_aio}
\alias{collect_aio_}
\title{Collect the Value of an Aio or List of Aios}
\usage{
collect_aio(x)

collect_aio_(x)
}
\arguments{
\item{x}{an Aio or list of Aios (objects of class \sQuote{sendAio},
\sQuote{recvAio} or \sQuote{ncurlAio}).}
}
\value{
Depending on the type of \sQuote{x} supplied, either an object or a
    list of objects (the same length as \sQuote{x}, preserving names).
}
\description{
\code{collect_aio} retrieves the value of a single, or list of, asynchronous
    Aio operations, waiting for their completion if still in progress.

\code{collect_aio_} is a variant that allows user interrupts, suitable for
    interactive use.
}
