% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_new}
\alias{cv_new}
\title{New Condition Variable}
\usage{
cv_new()
}
\value{
A 'conditionVariable' object.
}
\description{
Creates a new condition variable (protected by a mutex).
}
\details{
Pass the 'conditionVariable' object to the special \code{cv_} forms
    of the functions returning 'recvAio' objects. Completion of the receive,
    which happens asynchronously will signal the condition variable by
    incrementing it by 1.
}
\examples{
cv <- cv_new()

}
