% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_wait}
\alias{cv_wait}
\title{Wait Upon a Condition Variable}
\usage{
cv_wait(cv)
}
\arguments{
\item{cv}{a 'conditionVariable' object.}
}
\value{
Invisible NULL.
}
\description{
Waits upon a condition variable to be signalled by an incoming message being
    received.
}
\details{
The internal condition maintains a state. Each signal increments the
    condition by 1. Each time this function returns, the condition is
    decremented by 1.
}
\examples{
cv <- cv_new()
# cv_wait(cv) # uncommenting will block until the cv is signalled

}
