% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.interrupt}
\alias{.interrupt}
\title{Interrupt Switch}
\usage{
.interrupt(x = TRUE)
}
\arguments{
\item{x}{logical value.}
}
\value{
The logical value \code{x} supplied.
}
\description{
Sets whether async receive completions trigger an interrupt.
Internal package function.
}
\examples{
.interrupt()
.interrupt(FALSE)

}
\keyword{internal}
