% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mark}
\alias{.mark}
\title{Set Serialization Marker}
\usage{
.mark(x = FALSE)
}
\arguments{
\item{x}{logical value.}
}
\value{
Logical value 'x' supplied.
}
\description{
Internal package function.
}
\examples{
.mark(TRUE)
.mark()

}
\keyword{internal}
