% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mark}
\alias{.mark}
\title{Set Serialization Marker}
\usage{
.mark()
}
\value{
NULL.
}
\description{
Toggles the serialization marker on or off. Internal package function.
}
\examples{
.mark()
.mark()

}
\keyword{internal}
