% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mark}
\alias{.mark}
\title{Set Serialization Marker}
\usage{
.mark(x = TRUE)
}
\arguments{
\item{x}{logical value.}
}
\value{
The logical value \code{x} supplied.
}
\description{
Internal package function.
}
\examples{
.mark()
.mark(FALSE)

}
\keyword{internal}
