% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{.unresolved}
\alias{.unresolved}
\title{Technical Utility: Query if an Aio is Unresolved}
\usage{
.unresolved(x)
}
\arguments{
\item{x}{an Aio or list of Aios (objects of class 'sendAio', 'recvAio' or
'ncurlAio').}
}
\value{
Logical \code{TRUE} if \code{x} is an unresolved Aio or else \code{FALSE}, or if \code{x}
is a list, the integer number of unresolved Aios in the list.
}
\description{
Query whether an Aio or list of Aios remains unresolved. This is an
experimental technical utility version of \code{\link[=unresolved]{unresolved()}} not intended for
ordinary use. Provides a method of querying the busy status of an Aio without
altering its state in any way i.e. not attempting to retrieve the result or
message.
}
\details{
\code{.unresolved()} is not intended to be used for 'recvAio' returned by a
signalling function, in which case \code{\link[=unresolved]{unresolved()}} must be used in all cases.
}
\keyword{internal}
