% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{.unresolved}
\alias{.unresolved}
\title{Technical Utility: Query if an Aio is Unresolved}
\usage{
.unresolved(aio)
}
\arguments{
\item{aio}{an Aio (object of class 'sendAio' or 'recvAio').}
}
\value{
Logical TRUE if 'aio' is an unresolved Aio, or FALSE otherwise.
}
\description{
Query whether an Aio remains unresolved. This function is a technical utility
    version of \code{\link{unresolved}} not intended for ordinary use.
    Provides a method of querying the busy status of an Aio without altering
    its state in any way i.e. not attempting to retrieve the result or message.
}
