% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{\%~>\%}
\alias{\%~>\%}
\title{Signal Forwarder}
\usage{
cv \%~>\% cv2
}
\arguments{
\item{cv}{a 'conditionVariable' object, from which to forward the signal.}

\item{cv2}{a 'conditionVariable' object, to which to forward the signal.}
}
\value{
Invisibly, 'cv' (with the condition value reset to zero).
}
\description{
Forwards signals from one 'conditionVariable' to another.
}
\details{
This is an experimental operator.

    The condition value of 'cv' is reset to zero. It is intended that 'cv' be
    used as a counter and that waits should be performed on 'cv2'. Waiting on
    'cv' is not supported in conjunction with use of this operator.
}
\examples{
cv <- cv()
cv2 <- cv()

cv \%~>\% cv2

cv_signal(cv)
cv_value(cv)
cv_value(cv2)

}
