% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{\%~>\%}
\alias{\%~>\%}
\title{Signal Forwarder}
\usage{
cv \%~>\% cv2
}
\arguments{
\item{cv}{a 'conditionVariable' object, from which to forward the signal.}

\item{cv2}{a 'conditionVariable' object, to which to forward the signal.}
}
\value{
Invisibly, 'cv'.
}
\description{
Forwards signals from one 'conditionVariable' to another.
}
\details{
This is an experimental function. When used, waits should be
    performed only on 'cv2' and not 'cv'. Waiting on 'cv' is not supported.
}
\examples{
cv <- cv()
cv2 <- cv()

cv \%~>\% cv2

cv_signal(cv)
cv_value(cv)
cv_value(cv2)

}
