% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ip_addr}
\alias{ip_addr}
\title{IP Address}
\usage{
ip_addr()
}
\value{
A character string.
}
\description{
Returns a character string comprising the local network IPv4 address, or
vector if there are multiple addresses from multiple network adapters, or
an empty character string if unavailable.
}
\examples{
ip_addr()

}
