% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_aio}
\alias{is_aio}
\title{Is Aio}
\usage{
is_aio(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE if 'x' is of class 'recvAio' or 'sendAio', FALSE
    otherwise.
}
\description{
Is the object an Aio (sendAio or recvAio).
}
\examples{
sock <- socket(listen = "inproc://isaio")
r <- recv_aio(sock)
s <- send_aio(sock, "test")

is_aio(r)
is_aio(s)

close(sock)

}
