% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_aio}
\alias{is_aio}
\alias{is_nano}
\title{Validators}
\usage{
is_aio(x)

is_nano(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE or FALSE.
}
\description{
Validator functions for object types created by \{nanonext\}.
}
\details{
Is the object an Aio (inheriting from class \sQuote{sendAio} or
    \sQuote{recvAio}).

    Is the object an object inheriting from class \sQuote{nano} i.e. a
    nanoSocket, nanoContext, nanoStream, nanoListener, nanoDialer, or nano
    Object.
}
\examples{
sock <- socket(listen = "inproc://isaio")
r <- recv_aio(sock)
s <- send_aio(sock, "test")
is_aio(r)
is_aio(s)
close(sock)

s <- socket()
is_nano(s)
n <- nano()
is_nano(n)
close(s)
n$close()

}
