% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_error_value}
\alias{is_error_value}
\title{Is Error Value}
\usage{
is_error_value(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE if 'x' is of class 'errorValue', FALSE otherwise.
}
\description{
Is the object an error value generated by NNG. All receive functions class
    integer error codes as 'errorValue' to be easily distinguishable from
    integer message values.
}
\examples{
is_error_value(1L)

}
