% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_error_value}
\alias{is_error_value}
\title{Is Error Value}
\usage{
is_error_value(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE if 'x' is of class 'errorValue', FALSE otherwise.
}
\description{
Is the object an error value generated by NNG. All returned integer error
    codes are classed as 'errorValue' to be distinguishable from integer
    message values. Includes user-specified errors such as 'aio' timeouts.
}
\examples{
is_error_value(1L)

}
