% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_error_value}
\alias{is_error_value}
\title{Is Error Value}
\usage{
is_error_value(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE if 'x' is of class 'errorValue', FALSE otherwise.
}
\description{
Is the object an error value generated by NNG. All returned integer error
    codes are classed as 'errorValue' to be distinguishable from integer
    message values.
}
\section{Warnings}{


    A warning is generated every time an 'errorValue' is returned. The package
    changes the behaviour of warnings by setting \code{options(warn = 1)}
    upon package load. This prints the warning to the console as soon as it
    occurs, rather than the default of waiting for the top–level function to
    return.

    See \code{\link{warning}} or the entry for 'warn' in \code{\link{options}}
    for further details. The original option value for 'warn' is reverted upon
    package unload.
}

\examples{
is_error_value(1L)

}
