% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{is_resolved}
\alias{is_resolved}
\title{Is Resolved (Asynchronous AIO Operation)}
\usage{
is_resolved(aio)
}
\arguments{
\item{aio}{An Aio (object of class 'sendAio' or 'recvAio').}
}
\value{
Logical TRUE or FALSE. NA if 'aio' is not a 'sendAio' or 'recvAio'.
}
\description{
Query whether an Aio has resolved. This function is non-blocking unlike
    \code{\link{call_aio}} which waits for completion.
}
\details{
Querying resolution will potentially cause the state of the Aio to
    update.
}
\examples{
s1 <- socket("pair", listen = "inproc://nanonext")
aio <- send_aio(s1, "test", timeout = 100)
is_resolved(aio)

s2 <- socket("pair", dial = "inproc://nanonext")
is_resolved(aio)

close(s1)
close(s2)

}
