% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logging}
\alias{logging}
\title{Logging Level}
\usage{
logging(level = c("keep", "check", "error", "info"))
}
\arguments{
\item{level}{specify a logging level. No longer used.}
}
\value{
Invisible NULL. If the function is called with no arguments,
    the logical code of the logging level is returned instead.
}
\description{
This function is deprecated.
}
\details{
The environment variable 'NANONEXT_LOG' is checked automatically on
    package load. If the variable is set incorrectly, the default level
    of 'error' is used instead.
}
