% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logging}
\alias{logging}
\title{Logging Level}
\usage{
logging(level = c("error", "info"))
}
\arguments{
\item{level}{specify a logging level - either 'error' (default) which sends
all NNG errors to stderr or 'info' which also sends informational events
such as socket open, listener/dialer start, socket close etc. to stdout.}
}
\value{
Invisible NULL, or if 'level' is not specified, the integer code of
    the logging level.
}
\description{
Set the logging level of nanonext.
}
\examples{
logging(level = "info")
sock <- socket("respondent", dial = "inproc://nanolog")
logging(level = "error")
close(sock)

}
