% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nano_init}
\alias{nano_init}
\title{nanonext Initialise}
\usage{
nano_init(warn)
}
\arguments{
\item{warn}{Deprecated.}
}
\value{
Invisible NULL.
}
\description{
This function prints a deprecation message and no longer has any effect.
}
