% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncurl.R
\name{ncurl}
\alias{ncurl}
\title{ncurl}
\usage{
ncurl(
  url,
  convert = TRUE,
  follow = FALSE,
  method = NULL,
  headers = NULL,
  data = NULL,
  response = NULL,
  timeout = NULL,
  tls = NULL
)
}
\arguments{
\item{url}{the URL address.}

\item{convert}{[default TRUE] logical value whether to attempt conversion
of the received raw bytes to a character vector. Set to \code{FALSE} if
downloading non-text data.}

\item{follow}{[default FALSE] logical value whether to automatically follow
redirects (not applicable for async requests). If \code{FALSE}, the redirect
address is returned as response header 'Location'.}

\item{method}{(optional) the HTTP method as a character string. Defaults to
'GET' if not specified, and could also be 'POST', 'PUT' etc.}

\item{headers}{(optional) a named character vector or named list specifying
the HTTP request headers, for example: \cr
\code{c(Authorization = "Bearer APIKEY", "Content-Type" = "text/plain")} or \cr
\code{list(Authorization = "Bearer APIKEY", "Content-Type" = "text/plain")} \cr
An unnamed vector or list will be ignored.}

\item{data}{(optional) request data to be submitted. Must be a character
string or raw vector, and other objects are ignored. If a character vector,
only the first element is taken. When supplying binary data, the
appropriate 'Content-Type' header should be set to specify the binary
format.}

\item{response}{(optional) \code{TRUE} to return all response headers, or a
character vector specifying the response headers to return e.g.
\code{c("date", "server")}. These are case-insensitive and will return NULL if
not present. A non-character vector will be ignored (other than \code{TRUE}).}

\item{timeout}{(optional) integer value in milliseconds after which the
transaction times out if not yet complete.}

\item{tls}{(optional) applicable to secure HTTPS sites only, a client TLS
Configuration object created by \code{\link[=tls_config]{tls_config()}}. If missing or NULL,
certificates are not validated.}
}
\value{
Named list of 3 elements:
\itemize{
\item \verb{$status} - integer HTTP repsonse status code (200 - OK). Use
\code{\link[=status_code]{status_code()}} for a translation of the meaning.
\item \verb{$headers} - named list of response headers (all headers if
\code{response = TRUE}, or those specified in \code{response}, or NULL otherwise).
If the status code is within the 300 range, i.e. a redirect, the response
header 'Location' is automatically appended to return the redirect address.
\item \verb{$data} - the response body, as a character string if \code{convert = TRUE} (may
be further parsed as html, json, xml etc. as required), or a raw byte
vector if FALSE (use \code{\link[=writeBin]{writeBin()}} to save as a file).
}
}
\description{
nano cURL - a minimalist http(s) client.
}
\section{Public Internet HTTPS}{


When making HTTPS requests over the public internet, you should supply a TLS
configuration to validate server certificates. See \code{\link[=tls_config]{tls_config()}} for
details.
}

\examples{
ncurl(
  "https://postman-echo.com/get",
  response = TRUE,
  timeout = 2000L
)
ncurl(
  "https://postman-echo.com/put",
  method = "PUT",
  headers = c(Authorization = "Bearer APIKEY"),
  data = "hello world",
  timeout = 2000L
)
ncurl(
  "https://postman-echo.com/post",
  convert = FALSE,
  method = "POST",
  headers = c(`Content-Type` = "application/json"),
  data = '{"key":"value"}',
  response = c("date", "content-type"),
  timeout = 2000L
)

}
\seealso{
\code{\link[=ncurl_aio]{ncurl_aio()}} for asynchronous http requests; \code{\link[=ncurl_session]{ncurl_session()}} for
persistent connections.
}
