% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ncurl}
\alias{ncurl}
\title{ncurl}
\usage{
ncurl(http)
}
\arguments{
\item{http}{the URL/address of the resource to retrieve.}
}
\value{
Named list of 2 elements: 'raw' containing a raw vector of the received
    resource and 'data', the raw vector converted to a character string.
}
\description{
nano cURL - a minimalistic http(s) client.
}
\details{
Only performs HTTP GET operations, and does not follow HTTP redirects.

    The raw vector may be saved to a file using \code{\link{writeBin}} to
    re-create the original resource.

    The data vector is a character string allowing further parsing within R,
    as HTML, JSON etc.
}
\section{TLS Support}{


    Connecting to secure https sites is supported if your version of the NNG
    library was built with TLS support (using Mbed TLS) and the environment
    variable NANONEXT_TLS=1 was set when installing the package e.g. by
    \code{Sys.setenv(NANONEXT_TLS=1)}. Note: not applicable for Windows systems.
}

