% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{ncurl}
\alias{ncurl}
\title{ncurl}
\usage{
ncurl(
  url,
  async = FALSE,
  method = NULL,
  ctype = NULL,
  auth = NULL,
  data = NULL
)
}
\arguments{
\item{url}{the URL address.}

\item{async}{[default FALSE] logical value whether to perform actions async.}

\item{method}{(optional) the HTTP method.}

\item{ctype}{(optional) the 'Content-type' header.}

\item{auth}{(optional) the 'Authorization' header.}

\item{data}{(optional) the request data to be submitted.}
}
\value{
Named list of 2 elements:
    \itemize{
    \item{\code{$raw}} {- raw vector of the received resource (use
    \code{\link{writeBin}} to save to a file).}
    \item{\code{$data}} {- converted character string (if a recognised text
    format), or NULL otherwise. Other tools can be used to further parse this
    as html, json, xml etc. if required.}
    }

    Or else, if \code{async = TRUE}, a 'recvAio' (object of class 'recvAio').
}
\description{
nano cURL - a minimalist http(s) client.
}
\section{Redirects}{


    In interactive sessions: will prompt upon receiving a redirect location
    whether to follow or not (default: yes).

    In non-interactive sessions: redirects are never followed.
}

\section{TLS Support}{


    Connecting to secure https sites is supported if \code{\link{nng_version}}
    shows 'TLS supported'.
}

\examples{
ncurl("http://httpbin.org/get")
ncurl("http://httpbin.org/put", ,"PUT", "text/plain", "Bearer APIKEY", "hello world")
ncurl("http://httpbin.org/post", ,"POST", "application/json", ,'{"key": "value"}')

}
