% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncurl.R
\name{ncurl_session}
\alias{ncurl_session}
\alias{transact}
\title{ncurl Session}
\usage{
ncurl_session(
  url,
  convert = TRUE,
  method = NULL,
  headers = NULL,
  data = NULL,
  response = NULL,
  timeout = NULL,
  pem = NULL
)

transact(session)
}
\arguments{
\item{url}{the URL address.}

\item{convert}{[default TRUE] logical value whether to attempt conversion of
the received raw bytes to a character vector. Supplying a non-logical
value will error.}

\item{method}{(optional) the HTTP method (defaults to 'GET' if not specified).}

\item{headers}{(optional) a named list or character vector specifying the
HTTP request headers e.g. \code{list(`Content-Type` = "text/plain")} or
\code{c(Authorization = "Bearer APIKEY")}. Supplying a non-named list or
vector will error.}

\item{data}{(optional) the request data to be submitted.}

\item{response}{(optional) a character vector or list specifying the response
headers to return e.g. \code{c("date", "server")} or \code{list("Date", "Server")}.
These are case-insensitive and will return NULL if not present.}

\item{timeout}{(optional) integer value in milliseconds after which the
connection and subsequent transact attempts time out.}

\item{pem}{(optional) applicable to secure HTTPS sites only. The path to a
file containing X.509 certificate(s) in PEM format, comprising the
certificate authority certificate chain (and revocation list if present).
If missing or NULL, certificates are not validated.}

\item{session}{an 'ncurlSession' object.}
}
\value{
For \code{ncurl_session}: an 'ncurlSession' object if successful, or
    else an 'errorValue'.

For \code{transact}: a named list of 4 elements:
    \itemize{
    \item{\code{$status}} {- integer HTTP repsonse status code (200 - OK).
    Use \code{\link{status_code}} for a translation of the meaning.}
    \item{\code{$headers}} {- named list of response headers (if specified in
    the session), or NULL otherwise. If the status code is within the 300
    range, i.e. a redirect, the response header 'Location' is automatically
    appended to return the redirect address.}
    \item{\code{$raw}} {- raw vector of the received resource (use
    \code{\link{writeBin}} to save to a file).}
    \item{\code{$data}} {- converted character string (if specified in the
    session), or NULL otherwise. This may be further parsed this as html,
    json, xml etc. if required.}
    }
}
\description{
nano cURL - a minimalist http(s) client. A session encapsulates a connection,
    along with all related parameters, and may be used to return data
    multiple times by repeatedly calling \code{transact}, which transacts
    once over the connection.
}
\examples{
s <- ncurl_session("https://httpbin.org/get", response = "date", timeout = 5000L)
s
if (!is_error_value(s)) transact(s)
if (!is_error_value(s)) close(s)

}
