% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_config}
\alias{next_config}
\title{Configure Next Mode}
\usage{
next_config(inhook, outhook)
}
\arguments{
\item{inhook}{a function (for custom serialization). The signature for this
function must accept a list and return a raw vector, e.g.
\code{safetensors::safe_serialize}, or else NULL to reset.}

\item{outhook}{a function (for custom unserialization). The signature for
this function must accept a raw vector and return a list, e.g.
\code{safetensors::safe_load_file}, or else NULL to reset.}
}
\value{
Invisibly, a pairlist comprising the currently-registered 'next'
    configuration.
}
\description{
Configures 'next' mode by registering 'refhook' functions for serialization
    and unserialization. This permits sending and receiving reference
    objects, such as those accessed via an external pointer, between
    different R sessions.
}
\details{
Calling this function without any arguments returns (invisibly) the
    currently-registered 'next' configuration.

    Alternatively, calling this function with a configuration pairlist
    previously returned by this function registers the supplied configuration.
}
\section{Refhook}{


    The 'refhook' functions are a native feature of R's serialization
    mechanism and apply to all non-system reference objects (external
    pointers, weak references, and all environments other than namespace and
    package environments and the Global Environment).
}

\examples{
cfg <- next_config(inhook = function(x) serialize(x, NULL),
                   outhook = unserialize)
cfg

nul <- next_config(NULL, NULL)
print(next_config())

print(next_config(cfg))
print(next_config(nul))

}
