% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_config}
\alias{next_config}
\title{Configure Custom Serialization}
\usage{
next_config(refhook = list(), class = "", vec = FALSE, mark = FALSE)
}
\arguments{
\item{refhook}{not used.}

\item{class}{[default ""] not used.}

\item{vec}{[default FALSE] not used.}

\item{mark}{[default FALSE] (for advanced use only) logical value, whether to
mark serialized data with a special bit.}
}
\value{
NULL.
}
\description{
This function is defunct. Please refer to \link{serial_config} instead.
}
\keyword{internal}
