% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nextmode}
\alias{nextmode}
\title{Next Mode Settings}
\usage{
nextmode(inhook, outhook, mark = FALSE)
}
\arguments{
\item{inhook}{a function (for custom serialization). The signature for this
function must accept a list of external pointer objects and return a raw
vector, e.g. \code{torch::torch_serialize}, or else NULL to reset.}

\item{outhook}{a function (for custom unserialization). The signature for
this function must accept a raw vector and return a list of external
pointer objects, e.g. \code{torch::torch_load}, or else NULL to reset.}

\item{mark}{[default FALSE] (for advanced use only) logical value, whether to
mark serialized data with a special bit.}
}
\value{
Invisibly, a pairlist comprising the currently-registered 'inhook'
    and 'outhook' functions.
}
\description{
Configures send mode 'next' by registering functions for custom serialization
    and unserialization of external pointer objects, allowing these to be
    sent and received between different R sessions.
}
\details{
Calling this function without any arguments returns (invisibly) the
    currently-registered functions (and resets 'mark' to FALSE).
}
\examples{
cfg <- nextmode(inhook = function(x) serialize(x, NULL),
                outhook = unserialize,
                mark = TRUE)
cfg

nextmode(NULL, NULL)
print(nextmode())

}
