% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nng_error}
\alias{nng_error}
\title{Translate Error Codes}
\usage{
nng_error(xc)
}
\arguments{
\item{xc}{integer exit code to translate.}
}
\value{
A character vector.
}
\description{
Translate integer exit code to human-readable form. All package functions
    return an integer exit code on error rather than the expected return
    value. These are classed 'errorValue' and may be checked by the function
    \code{\link{is_error_value}}.
}
\examples{
nng_error(1L)

}
