% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nng_version}
\alias{nng_version}
\title{NNG Library Version}
\usage{
nng_version()
}
\value{
A character vector of length 2.
}
\description{
Returns the version of 'libnng' used and whether TLS is supported.
}
\section{TLS Support}{


    If a system installation of Mbed TLS ‘libmbedtls’ is detected in the
    standard filesystem locations, NNG will link against this and be built
    with TLS support.

    The environment variable Sys.setenv(NANONEXT_TLS=1) may also be set prior
    to installation to enable TLS support if ‘libmbedtls’ is installed in a
    non-standard location.

    Note: this is not applicable to Windows systems.
}

\examples{
nng_version()

}
