% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nng_version}
\alias{nng_version}
\title{NNG Library Version}
\usage{
nng_version()
}
\value{
A character vector of length 2.
}
\description{
Returns the versions of the \sQuote{libnng} and \sQuote{libmbedtls} libraries
used by the package.
}
\examples{
nng_version()

}
