% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nng_version}
\alias{nng_version}
\title{NNG Library Version}
\usage{
nng_version()
}
\value{
A character vector of length 2.
}
\description{
Returns the version of 'libnng' used and whether TLS is supported.
}
\section{TLS Support}{


    Where system installations of 'libnng' and 'libmbedtls' development
    headers are detected in the same location, it is assumed that NNG was
    built with TLS support (using Mbed TLS) and TLS is configured appropriately.

    Otherwise, the environment variable `Sys.setenv(NANONEXT_TLS=1)` may be
    set prior to installation if:

    - your system installations of 'libnng' (built with TLS support) and
    'libmbedtls' are in different locations; or

    - you have a system installation of 'libmbedtls' but not 'libnng' and want
    nanonext to download and build a more recent version of 'libnng' than
    available in system repositories against this.

    Note: this is not applicable to Windows systems.
}

\examples{
nng_version()

}
