% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipe_notify}
\alias{pipe_notify}
\title{Pipe Notify}
\usage{
pipe_notify(socket, cv, open = TRUE, close = TRUE)
}
\arguments{
\item{socket}{a Socket.}

\item{cv}{a 'conditionVariable' object.}

\item{open}{[default TRUE] logical value whether to signal on pipe open.}

\item{close}{[default TRUE] logical value whether to signal on pipe close.}
}
\value{
Invisibly, an integer exit code (zero on success).
}
\description{
Signals a 'conditionVariable' whenever pipes are created or closed at a socket.
}
\details{
For pipe open: this event occurs after the pipe is fully added to
    the socket. Prior to this time, it is not possible to communicate over
    the pipe with the socket.

    For pipe close: this event occurs after the pipe has been removed from
    the socket. The underlying transport may be closed at this point, and it
    is not possible communicate using this pipe.
}
\examples{
s <- socket(listen = "inproc://nanopipe")
cv <- cv()
pipe_notify(s, cv, open = TRUE, close = TRUE)
cv_value(cv)

s1 <- socket(dial = "inproc://nanopipe")
cv_value(cv)
close(s1)
cv_value(cv)

close(s)

}
