% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random}
\alias{random}
\title{Random Number Generator}
\usage{
random(n = 1L)
}
\arguments{
\item{n}{[default 1L] length of vector to return.}
}
\value{
A length 'n' vector of random doubles.
}
\description{
Returns a vector of random numbers of the length specified.
}
