% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random}
\alias{random}
\title{NNG Random Number Generator}
\usage{
random(n = 1L)
}
\arguments{
\item{n}{[default 1L] length of vector to return.}
}
\value{
A length 'n' vector of random positive doubles. Each random number
    has a range between 0 and 2 * \code{.Machine$integer.max} inclusive.
}
\description{
Strictly not for statistical analysis. Not reproducible. No ability to set
    a seed value. Provides random numbers suitable for system functions such
    as cryptographic key generation. Random values are obtained using
    platform-specific strong cryptographic random number facilities where
    available.
}
\details{
If 'n' is non-integer, it will be coerced to integer; if a vector,
    only the first element will be used.
}
\examples{
random()
random(n = 3L)

}
