% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random}
\alias{random}
\title{Random Data Generation}
\usage{
random(n = 1L)
}
\arguments{
\item{n}{[default 1L] integer number of random byte of data to generate.}
}
\value{
A length one vector of '2n' random characters.
}
\description{
Strictly not for use in statistical analysis. Non-reproducible and with
    unknown statistical properties. Provides an alternative source of
    randomness from the Mbed TLS library using its own entropy collection,
    for purposes such as cryptographic key generation.
}
\details{
If 'n' is non-integer, it will be coerced to integer; if a vector,
    only the first element will be used.

    The random bytes of data generated are represented as hex values,
    producing 2 characters for each byte.
}
\note{
Results obtained are independent of and do not alter the state of R's
    own pseudo-random number generators.
}
\examples{
random()
random(8L)

}
