% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random}
\alias{random}
\title{NNG Random Number Generator}
\usage{
random()
}
\value{
A (positive) double.
}
\description{
Strictly not for statistical analysis. Not reproducible as no ability to set
    a seed value. Provides a random number suitable for system functions such
    as cryptographic key generation. The value is obtained using
    platform-specific strong cryptographic random number facilities where
    available.
}
\examples{
random()

}
