% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random}
\alias{random}
\title{NNG Random Number Generator}
\usage{
random(n = 1L)
}
\arguments{
\item{n}{[default 1L] length of vector to return (if non-integer, will be
coerced to integer).}
}
\value{
A length 'n' vector of random positive doubles.
}
\description{
Strictly not for statistical analysis. Not reproducible as no ability to set
    a seed value. Provides random numbers suitable for system functions such
    as cryptographic key generation. Random values are obtained using
    platform-specific strong cryptographic random number facilities where
    available.
}
\examples{
random()
random(n = 3L)

}
