% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_stdin}
\alias{read_stdin}
\title{Read stdin}
\usage{
read_stdin()
}
\value{
a Socket.
}
\description{
Read \code{stdin} from a background thread, allowing the stream to be accessed as
messages from an NNG 'inproc' socket. As the read is blocking, it can only be
used in non-interactive sessions. This function is experimental.
}
