% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{reap}
\alias{reap}
\title{Reap}
\usage{
reap(con)
}
\arguments{
\item{con}{a Socket, Context, Listener or Dialer.}
}
\value{
Integer zero.
}
\description{
An alternative to \code{close} for Sockets, Contexts, Listeners and Dialers
    that avoids S3 method dispatch. Intended to be called prior to objects
    being discarded - do not attempt to re-use an object supplied to this
    function.
}
\details{
May be used on unclassed external pointers e.g. those created by
    \code{\link{.context}}. For listeners, runs the finalizer and cleans up
    associated resources including releasing the bound address. Returns
    silently and does not warn or error, nor does it update the state of
    object attributes.
}
\examples{
s <- socket("req")
listen(s)
dial(s)
ctx <- .context(s)

reap(ctx)
reap(s[["dialer"]][[1]])
reap(s[["listener"]][[1]])
reap(s)

}
\seealso{
\code{\link{close}}
}
