% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{reap}
\alias{reap}
\title{Reap}
\usage{
reap(con)
}
\arguments{
\item{con}{a Socket or Context.}
}
\value{
Invisibly, an integer exit code (zero on success).
}
\description{
A faster alternative to \code{close} for Sockets and Contexts avoiding S3
    method dispatch.
}
\details{
May be used on unclassed external pointers e.g. those created by
    \code{\link{.context}}.
}
\examples{
s <- socket("req")
ctx <- .context(s)

reap(ctx)
reap(s)

}
\seealso{
\code{\link{close}}
}
