% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{recover_aio}
\alias{recover_aio}
\title{Recover an Aio Value}
\usage{
recover_aio(aio)
}
\arguments{
\item{aio}{an ncurlAio (object of class 'ncurlAio').}
}
\value{
The passed object (invisibly).
}
\description{
For use on 'ncurlAio' when encountering a character conversion error
    'embedded nul in string' (received data is not text data). This function
    allows recovery by retrieving the value as a raw vector.
}
\details{
A raw vector will be available at \code{$data}, as if
    \code{convert = FALSE} had been specified.

    Note: calling this function on an unresolved Aio is an error and any
    resulting operation is not guaranteed.
}
\examples{
res <- ncurl_aio("https://shikokuchuo.net/nanonext/reference/figures/logo.png")
tryCatch(call_aio(res), error = identity)
recover_aio(res)$data

}
