% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{refhook}
\alias{refhook}
\title{Refhook Constructor}
\usage{
refhook(valid, encode, decode)
}
\arguments{
\item{valid}{validation code e.g. \code{inherits(x, "torch_tensor")} or
\code{typeof(x) == "externalptr"}.}

\item{encode}{the encode function returning a raw vector e.g. \code{torch:::torch_serialize(x)}.}

\item{decode}{the decode function from a raw vector back to the original
object e.g. \code{torch::torch_load(x)}.}
}
\value{
A function.
}
\description{
Create a function with the required signature for the 'refhook' argument of
    \code{\link{socket}}.
}
\details{
The expressions supplied should all reference the single argument
    '\code{x}'.

    'valid' will be evaluated within an \code{if()} block and 'encode' and
    'decode' should be wrapped in \code{{}} if a compound statement.
}
\examples{
refhook(typeof(x) == "weakref", base64enc(weakref_value(x), convert = FALSE), base64dec(x))

}
