% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refhook}
\alias{refhook}
\title{Set and Retrieve Refhook Function}
\usage{
refhook(x = list())
}
\arguments{
\item{x}{a function or base64 encoded function, or NULL to reset.}
}
\value{
The current refhook function stored internally.
}
\description{
Providing an argument sets the refhook function stored internally, used as
    the argument 'refhook' in all subsequent \code{\link{serialize}} and
    \code{\link{unserialize}} operations. Calling without an argument
    retrieves the value stored.
}
\examples{
refhook()

}
