% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socket.R
\name{refhook}
\alias{refhook}
\title{Refhook Constructor}
\usage{
refhook(valid, encode, decode)
}
\arguments{
\item{valid}{an expression of the validation e.g. \code{inherits(x, "torch_tensor")}
or \code{typeof(x) == "externalptr"}.}

\item{encode}{an expression of the encode function e.g. \code{weakref_value(x)}.}

\item{decode}{an expression of the decode function e.g. \code{unserialize(x)}.}
}
\value{
a function.
}
\description{
Create a function with the required signature for the 'refhook' argument of
    \code{\link{socket}}.
}
\details{
The expressions supplied should all reference the single argument
    '\code{x}'.

    'valid' will be evaluated within an \code{if()} block and 'encode' and
    'decode' should be wrapped in \code{{}} if a compound statement.
}
\examples{
refhook(typeof(x) == "weakref", weakref_value(x), x)

}
