% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serial_config}
\alias{serial_config}
\title{Configure Custom Serialization}
\usage{
serial_config(class, sfunc = NULL, ufunc = NULL, vec = FALSE)
}
\arguments{
\item{class}{character string of the class of object custom serialization
functions are applied to, e.g. \sQuote{ArrowTabular} or
\sQuote{torch_tensor}, or else NULL to reset.}

\item{sfunc}{a function that accepts a reference object inheriting from
\sQuote{class} (or a list of such objects) and returns a raw vector.}

\item{ufunc}{a function that accepts a raw vector and returns a reference
object (or list of such objects).}

\item{vec}{[default FALSE] whether or not the serialization functions are
vectorized and accept and return a list of reference objects, e.g.
\code{torch::torch_serialize} and \code{torch::torch_load}, or if FALSE
return reference object individually e.g. \code{arrow::write_to_raw} and
\code{arrow::read_ipc_stream}.}
}
\value{
A pairlist comprising the currently-registered configuration.
}
\description{
Registers functions on a Socket for custom serialization and unserialization
    of non-system reference objects, allowing these to be sent and received
    between different R sessions. Registered functions apply to all send and
    receive operations in mode \sQuote{serial} performed over the Socket,
    including those using a Context.
}
\examples{
serial_config("test_cls", function(x) serialize(x, NULL), unserialize)

}
