% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{set_promise_context}
\alias{set_promise_context}
\title{Set Promise Context}
\usage{
set_promise_context(x, ctx)
}
\arguments{
\item{x}{an 'recvAio' object returned by \code{\link{request2}}.}

\item{ctx}{the context environment.}
}
\value{
The object 'x'.
}
\description{
When called in the correct context on a 'recvAio' object returned by
    \code{\link{request2}}, creates an event-driven promise that will resolve
    asynchronously when the request is complete.
}
\details{
The object passed as 'x' is returned regardless of whether the
    promise context was set successfully or not. If successful, 'x' is
    modified in place with the promise context.
}
