% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{setopt}
\alias{setopt}
\alias{setopt.nanoSocket}
\alias{setopt.nanoDialer}
\alias{setopt.nanoListener}
\alias{setopt.nanoContext}
\title{Set Option on Socket, Context, Dialer or Listener}
\usage{
setopt(
  object,
  type = c("bool", "int", "ms", "size", "string", "uint64"),
  opt,
  value
)

\method{setopt}{nanoSocket}(
  object,
  type = c("bool", "int", "ms", "size", "string", "uint64"),
  opt,
  value
)

\method{setopt}{nanoDialer}(
  object,
  type = c("bool", "int", "ms", "size", "string", "uint64"),
  opt,
  value
)

\method{setopt}{nanoListener}(
  object,
  type = c("bool", "int", "ms", "size", "string", "uint64"),
  opt,
  value
)

\method{setopt}{nanoContext}(
  object,
  type = c("bool", "int", "ms", "size", "string", "uint64"),
  opt,
  value
)
}
\arguments{
\item{object}{a Socket, Context, Listener or Dialer.}

\item{type}{[default 'bool'] type of option - either 'bool', 'int', 'ms'
(duration), 'size', 'string' or 'uint64'.}

\item{opt}{name of option, e.g. 'reconnect-time-min', as a character string.
See \link{options}.}

\item{value}{value of option.}
}
\value{
Zero (invisibly) on success.
}
\description{
Set \link{options} on a Socket, Context, Dialer or Listener.
}
\details{
Note: once a dialer or listener has started, it is not generally
    possible to change its configuration. Hence create the dialer or listener
    with 'autostart = FALSE' if configuration needs to be set.

    To set options on a Listener or Dialer attached to a Socket or nano object,
    you must pass in the objects directly via for example \code{$listener[[1]]}
    for the first Listener.
}
\examples{
s <- socket("pair")
setopt(s, "ms", "recv-timeout", 2000)
close(s)

s <- socket("pair", dial = "inproc://nanonext", autostart = FALSE)
setopt(s$dialer[[1]], "ms", "reconnect-time-min", 2000)
start(s$dialer[[1]])
close(s)

s <- socket("pair", listen = "inproc://nanonext", autostart = FALSE)
setopt(s$listener[[1]], "size", "recv-size-max", 1024)
start(s$listener[[1]])
close(s)

s <- socket("req")
ctx <- context(s)
setopt(ctx, "ms", "send-timeout", 2000)
close(ctx)
close(s)

}
