% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha.R
\name{sha256}
\alias{sha256}
\title{Cryptographic Hashing Using the SHA256 Algorithm}
\usage{
sha256(x, key = NULL)
}
\arguments{
\item{x}{an object.}

\item{key}{[default NULL] optionally supply a secret key to generate an HMAC.
Returns the SHA256 hash of 'x' if missing or NULL.}
}
\value{
A 'nanoHash' object - a raw vector of 32 bytes.
}
\description{
Returns a SHA256 hash or HMAC of the supplied R object. Uses the optimised
    implementation from the Mbed TLS library.
}
\details{
For arguments 'x' and 'key', a raw vector is hashed directly, a
    character string is converted using \code{\link{charToRaw}}, whilst other
    objects are serialised first.
}
\examples{
sha256("hello world!")

# Converts to a single character string:
as.character(sha256("hello world!"))

# Obtain HMAC:
sha256("hello world!", "SECRET_KEY")

}
