% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha.R
\name{sha256}
\alias{sha256}
\title{Cryptographic Hashing Using the sha256 Algorithm}
\usage{
sha256(x)
}
\arguments{
\item{x}{an object. A raw vector or character string will be hashed directly
whilst other objects are serialised first.}
}
\value{
A raw vector of 32 bytes.
}
\description{
Returns a sha256 hash of the supplied R object.
}
\details{
Hashing capabilities rely on the 'mbedTLS' library. If the package
    was not linked against 'mbedTLS' at install time, an errorValue 9 'not
    supported' will be returned instead.
}
\examples{
sha256("hello world!")

# To contatenate the hash into a single character string:
paste(sha256("hello world!"), collapse = "")

}
