% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listdial.R
\name{start}
\alias{start}
\alias{start.nanoListener}
\alias{start.nanoDialer}
\title{Start Listener/Dialer}
\usage{
\method{start}{nanoListener}(x, ...)

\method{start}{nanoDialer}(x, async = TRUE, ...)
}
\arguments{
\item{x}{a Listener or Dialer.}

\item{...}{not used.}

\item{async}{[default TRUE] (applicable to Dialers only) logical flag whether
the connection attempt, including any name resolution, is to be made
asynchronously. This behaviour is more resilient, but also generally makes
diagnosing failures somewhat more difficult. If FALSE, failure, such as if
the connection is refused, will be returned immediately, and no further
action will be taken.}
}
\value{
Invisibly, an integer exit code (zero on success).
}
\description{
Start a Listener/Dialer.
}
