% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{status_code}
\alias{status_code}
\title{Translate HTTP Status Codes}
\usage{
status_code(x)
}
\arguments{
\item{x}{numeric HTTP status code to translate.}
}
\value{
A character vector comprising the status code and explanation
separated by \sQuote{ | }.
}
\description{
Provides an explanation for HTTP response status codes (in the range 100 to
599). If the status code is not defined as per RFC 9110,
\sQuote{Unknown HTTP Status} is returned - this may be a custom code used by
the server.
}
\examples{
status_code(200)
status_code(404)

}
