% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{stop_aio}
\alias{stop_aio}
\title{Stop Asynchronous Aio Operation}
\usage{
stop_aio(x)
}
\arguments{
\item{x}{an Aio or list of Aios (objects of class \sQuote{sendAio},
\sQuote{recvAio} or \sQuote{ncurlAio}).}
}
\value{
Invisible NULL.
}
\description{
Stop an asynchronous Aio operation, or a list of Aio operations.
}
\details{
Stops the asynchronous I/O operation associated with \sQuote{aio} by
    aborting, and then waits for it to complete or to be completely aborted,
    and for the callback associated with the \sQuote{aio} to have completed
    executing. If successful, the \sQuote{aio} will resolve to an
    \sQuote{errorValue} 20 (Operation canceled).

    Note this function operates silently and does not error even if
    \sQuote{aio} is not an active Aio, always returning invisible NULL.
}
