% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{stop_aio}
\alias{stop_aio}
\title{Stop Asynchronous Aio Operation}
\usage{
stop_aio(aio)
}
\arguments{
\item{aio}{An Aio (object of class 'sendAio' or 'recvAio').}
}
\value{
Invisible NULL.
}
\description{
Stop an asynchronous Aio operation.
}
\details{
Stops the asynchronous I/O operation associated with 'aio' by
    aborting, and then waits for it to complete or to be completely aborted.
    The Aio is then deallocated and no further operations may be performed on
    it.

    Note this function operates silently and does not error even if 'aio' is
    not an active Aio, always returning invisible NULL.
}
