% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{stop_aio}
\alias{stop_aio}
\title{Stop Asynchronous AIO Operation}
\usage{
stop_aio(aio)
}
\arguments{
\item{aio}{An Aio (object of class 'sendAio' or 'recvAio').}
}
\value{
Invisible NULL.
}
\description{
Stop an asynchronous AIO operation.
}
\details{
Stops the asynchronous I/O operation associated with 'aio' by
    aborting, and then waits for it to complete or to be completely aborted.
    The Aio is then deallocated and no further operations may be performed on
    it.
}
