% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strcat}
\alias{strcat}
\title{Concatenate Strings}
\usage{
strcat(a, b)
}
\arguments{
\item{a}{character value.}

\item{b}{character value.}
}
\value{
A character string.
}
\description{
A fast implementation that combines two character values into a single string.
    This function is deprecated and will be removed in a future version.
}
\details{
If either \sQuote{a} or \sQuote{b} is a vector of length greater
    than 1, only the first element of each is concatenated.
}
\examples{
strcat("hello ", "world!")

}
\keyword{internal}
