% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream_send}
\alias{stream_send}
\title{Send via Stream}
\usage{
stream_send(stream, data, timeout, echo = TRUE)
}
\arguments{
\item{stream}{a Stream.}

\item{data}{a vector.}

\item{timeout}{(optional) in milliseconds.}

\item{echo}{[default TRUE] logical flag whether to return the raw vector of
sent data. Set to FALSE for performance-critical applications where zero
will be returned (invisibly) instead.}
}
\value{
Raw vector of sent data, or (invisibly) an integer exit code (zero on
    success) if 'echo' is set to FALSE.
}
\description{
Send data over a byte stream. This is a low-level interface intended for
    communicating with non-NNG endpoints [experimental].
}
\details{
Sending a byte stream synchronously will block if the send is in
    progress and has not yet completed. Set a timeout to ensure the function
    returns under all scenarios.
}
