% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{subscribe}
\alias{subscribe}
\title{Subscribe Topic}
\usage{
subscribe(socket, topic = NULL)
}
\arguments{
\item{socket}{a Socket using the sub protocol.}

\item{topic}{[default NULL] a topic (vector or NULL). The default NULL
subscribes to all topics.}
}
\value{
Invisibly, an integer exit code (zero on success).
}
\description{
For a socket using the sub protocol in a publisher/subscriber pattern. Set a
    topic to subscribe to.
}
\details{
To use pub/sub the publisher must:
    \itemize{
    \item{specify \code{mode = 'raw'} when sending to allow the topics to be
    recognised by the receiving party.}
    \item{when sending a character vector, the topic must be separated from
    the rest of the message using e.g.
    \code{send(socket, c("topic", "message"), mode = "raw")}}
    }
}
\examples{
pub <- socket("pub", listen = "inproc://nanonext")
sub <- socket("sub", dial = "inproc://nanonext")

subscribe(sub, "examples")
send(pub, c("examples", "this is an example"), mode = "raw")
recv(sub, "character")
send(pub, c("other", "this other topic will not be received"), mode = "raw")
recv(sub, "character")

close(pub)
close(sub)

}
