% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncurl.R
\name{tls_config}
\alias{tls_config}
\title{TLS - Create Configuration}
\usage{
tls_config(file, server = FALSE, auth = TRUE)
}
\arguments{
\item{file}{absolute path to a file containing X.509 certificate(s) in PEM
format, comprising the certificate authority certificate chain (and
revocation list if present).}

\item{server}{[default FALSE] logical value whether to configure for a server
or else a client.}

\item{auth}{[default TRUE] logical value whether to require authentication,
in which case a check is made to ensure that the peer has presented a
valid certificate. If the certificate is invalid or missing, then the
session is refused. Otherwise, authentication is optional, in which case
a certificate is validated if presented by the peer, but if not then the
session is allowed to proceed without authentication.}
}
\value{
A 'tlsConfig' object.
}
\description{
Create a TLS configuration object to be used for secure connections.
}
\details{
Up-to-date CA certificates in PEM format, extracted from Mozilla,
    are available at the following URL:
    \url{https://curl.se/docs/caextract.html}.
}
