% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{tls_config}
\alias{tls_config}
\title{Create TLS Configuration}
\usage{
tls_config(client = NULL, server = NULL, pass = NULL, auth = TRUE)
}
\arguments{
\item{client}{(optional) for creating a client configuration, the absolute
path to a file containing X.509 certificate(s) in PEM format, comprising
the certificate authority certificate chain (and revocation list if
present), used to validate certificates presented by peers.}

\item{server}{(optional) for creating a server configuration, absolute path
to a single file containing the PEM encoded certificate and associated
private key, and any additional certificates leading to a validation
chain, with the leaf certificate first. There is no need to include the
self-signed root.}

\item{pass}{[default NULL] required only if the secret key contained in the
file supplied to 'server' is encrypted with a password. Do not provide
directly in case it is cached or recorded in history, but through an
object or function that returns the value.}

\item{auth}{[default TRUE] logical value whether to require authentication,
in which case a check is made to ensure that the peer has presented a
valid certificate. If the certificate is invalid or missing, then the
session is refused. Otherwise, authentication is optional, in which case
a certificate is validated if presented by the peer, but if not then the
session is allowed to proceed without authentication. If both 'client'
and 'server' are not supplied, then no authentication is performed and
this argument has no effect. Supplying a non-logical value will error.}
}
\value{
A 'tlsConfig' object.
}
\description{
Create a TLS configuration object to be used for secure connections.
}
\details{
Up-to-date CA certificates in PEM format, extracted from Mozilla,
    are available at: \url{https://curl.se/docs/caextract.html}. This link is
    not endorsed; use at your own risk.
}
\examples{
cfg <- tls_config()
ncurl("https://www.r-project.org/", timeout = 1000L, secure = cfg)

}
