% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aio.R
\name{unresolved}
\alias{unresolved}
\title{Query if an Aio is Unresolved}
\usage{
unresolved(aio)
}
\arguments{
\item{aio}{An Aio (object of class 'sendAio' or 'recvAio'), or Aio value
stored in \code{$result}, \code{$raw} or \code{$data} as the case may be.}
}
\value{
Logical TRUE or FALSE.
}
\description{
Query whether an Aio or Aio value remains unresolved. Unlike
    \code{\link{call_aio}}, this function does not wait for completion.
}
\details{
Returns TRUE for unresolved Aios or Aio values, FALSE otherwise.
    Suitable for use in control flow statements such as \code{while} or \code{if}.
}
\examples{
s1 <- socket("pair", listen = "inproc://nanonext")
aio <- send_aio(s1, "test", timeout = 100)

while (unresolved(aio)) {
  # do stuff before checking resolution again
  cat("unresolved\n")
  s2 <- socket("pair", dial = "inproc://nanonext")
  Sys.sleep(0.01)
}

unresolved(aio)

close(s1)
close(s2)

}
