% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{unsubscribe}
\alias{unsubscribe}
\alias{unsubscribe.nanoSocket}
\alias{unsubscribe.nanoContext}
\title{Unsubscribe Topic}
\usage{
unsubscribe(con, topic = NULL)

\method{unsubscribe}{nanoSocket}(con, topic = NULL)

\method{unsubscribe}{nanoContext}(con, topic = NULL)
}
\arguments{
\item{con}{a Socket or Context using the 'sub' protocol.}

\item{topic}{[default NULL] an atomic type or NULL. The default NULL
unsubscribes from all topics (if all topics were previously subscribed).}
}
\value{
Invisibly, an integer exit code (zero on success).
}
\description{
For a socket or context using the sub protocol in a publisher/subscriber
    pattern. Remove a topic from the subscription list.
}
\details{
Note that if the topic was not previously subscribed to then an
    'entry not found' error will result.

    To use pub/sub the publisher must:
    \itemize{
    \item{specify \code{mode = 'raw'} when sending to allow the topics to be
    recognised by the receiving party.}
    \item{when sending a character vector, the topic must be separated from
    the rest of the message using e.g.
    \code{send(socket, c("topic", "message"), mode = "raw")}}
    }
}
\examples{
pub <- socket("pub", listen = "inproc://nanonext")
sub <- socket("sub", dial = "inproc://nanonext")

subscribe(sub, NULL)
send(pub, c("examples", "this is an example"), mode = "raw")
recv(sub, "character")
unsubscribe(sub, NULL)
send(pub, c("examples", "this example will not be received"), mode = "raw")
recv(sub, "character")

subscribe(sub, 2)
send(pub, c(2, 10, 10, 20), mode = "raw")
recv(sub, "double", keep.raw = FALSE)

close(pub)
close(sub)

}
