% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_stdout}
\alias{write_stdout}
\title{Write to Stdout}
\usage{
write_stdout(x)
}
\arguments{
\item{x}{character string.}
}
\value{
Invisible NULL. As a side effect, \code{x} is output to \code{stdout}.
}
\description{
Performs a non-buffered write to \code{stdout} using the C function \code{write()}.
Avoids interleaved output when writing concurrently from multiple processes.
}
\details{
This function writes to the C-level \code{stdout} of the process and hence cannot
be re-directed by \code{\link[=sink]{sink()}}.
}
\examples{
write_stdout("\n")

}
