% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zzz}
\alias{zzz}
\title{Internal Package Function}
\usage{
zzz()
}
\description{
Only present for cleaning up after running examples and tests. Do not attempt
    to run the examples.
}
\examples{
rm(list = ls())
gc()
Sys.sleep(1L)
.Call(nanonext:::rnng_fini)

}
\keyword{internal}
